% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMSY.R
\name{addRecruitmentCurve}
\alias{addRecruitmentCurve}
\alias{addRecruitmentCurve.sam}
\title{Add stock-recruitment curve to srplot}
\usage{
addRecruitmentCurve(fit, CI = TRUE, col = rgb(0.6, 0, 0),
  cicol = rgb(0.6, 0, 0, 0.3), ...)

\method{addRecruitmentCurve}{sam}(fit, CI = TRUE, col = rgb(0.6, 0, 0),
  cicol = rgb(0.6, 0, 0, 0.3), ...)
}
\arguments{
\item{fit}{Object to show SR-curve for}

\item{CI}{Add confidence intervals?}

\item{col}{Color of fitted line}

\item{cicol}{Color of confidence intervals}

\item{...}{not used}
}
\description{
Add stock-recruitment curve to srplot
}
\seealso{
srplot
}
\author{
Christoffer Moesgaard Albertsen
}
