% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{b0plot}
\alias{b0plot}
\alias{b0plot.default}
\alias{b0plot.samforecast}
\alias{b0plot.hcr}
\title{SAM equilibrium biomass in the absence of fishing plot}
\usage{
b0plot(fit, ...)

\method{b0plot}{default}(fit, ...)

\method{b0plot}{samforecast}(fit, ...)

\method{b0plot}{hcr}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM equilibrium biomass in the absence of fishing plot
}
\details{
Plot of deterministic equilibrium biomass in the absence of fishing assuming biological parameters and selectivity for that year remains unchanged in the future.
}
