% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{b0table}
\alias{b0table}
\alias{b0table.default}
\title{B0 biomass table}
\usage{
b0table(fit, ...)

\method{b0table}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
B0 biomass table
}
\details{
...
}
