% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.sam}
\alias{c.sam}
\title{Collect sam objects}
\usage{
\method{c}{sam}(...)
}
\arguments{
\item{...}{one or more sam fits (as returned from the \code{\link{sam.fit}} function) to be combined}
}
\description{
Collect sam objects
}
\details{
...
}
