% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{catchbyfleetplot}
\alias{catchbyfleetplot}
\title{SAM catchbyfleet plot}
\usage{
catchbyfleetplot(fit, obs.show = FALSE, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{obs.show}{if observations are to be shown also}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM catchbyfleet plot
}
\details{
Plot of estimated (and optionally observed) total catch in weight
}
