% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{caytable}
\alias{caytable}
\title{Catch-at-age in numbers table}
\usage{
caytable(fit, fleet = which(fit$data$fleetTypes == 0))
}
\arguments{
\item{fit}{a fitted object of class 'sam' as returned from sam.fit}

\item{fleet}{the fleet number(s) to return catch summed for (default is to return the sum of all residual fleets).}
}
\description{
Catch-at-age in numbers table
}
\details{
...
}
