% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{clean.void.catches}
\alias{clean.void.catches}
\title{remove void catches}
\usage{
clean.void.catches(dat, conf)
}
\arguments{
\item{dat}{data for the sam model as returned from the setup.sam.data function}

\item{conf}{model configuration which can be set up using the \code{\link{defcon}} function and then modified}
}
\value{
an updated dataset without the catches where F is fixed to zero
}
\description{
remove void catches
}
