% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.sam}
\alias{coef.sam}
\title{Extract fixed coefficients of sam object}
\usage{
\method{coef}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object as returned from the \code{\link{sam.fit}} function}

\item{...}{extra arguments}
}
\description{
Extract fixed coefficients of sam object
}
\details{
fixed coefficients of sam object
}
