% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{componentplot}
\alias{componentplot}
\alias{componentplot.sam}
\title{Area plot of spawning components}
\usage{
componentplot(fit, ...)

\method{componentplot}{sam}(
  fit,
  onlyComponentYears = FALSE,
  ylab = "Composition",
  colSet = c("#332288", "#88CCEE", "#44AA99", "#117733", "#999933", "#DDCC77", "#661100",
    "#CC6677", "#882255", "#AA4499"),
  legend.pos = "bottom",
  bg = "white",
  ncol = length(cf),
  ...
)
}
\arguments{
\item{fit}{sam fit}

\item{...}{passed to legend}

\item{onlyComponentYears}{If true, x axis is limited to the range with spawning component data. Otherwise, the model years are used.}

\item{ylab}{Label for y axis}

\item{colSet}{Colors}

\item{legend.pos}{Legend position. See ?legend}

\item{bg}{Background of legend. See ?legend}

\item{ncol}{Number of columns in legend. See ?legend}
}
\value{
Nothing
}
\description{
Area plot of spawning components
}
\author{
Christoffer Moesgaard Albertsen
}
