% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{corplotcommon}
\alias{corplotcommon}
\title{Common function for plotting correlation matrices.}
\usage{
corplotcommon(x, fn, ...)
}
\arguments{
\item{x}{a list of correlation matrices}

\item{fn}{a vector of fleet names}

\item{...}{extra arguments to plotcorr}
}
\description{
Common function for plotting correlation matrices.
}
