% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defcon}
\alias{defcon}
\title{Setup basic minimal configuration for sam assessment}
\usage{
defcon(dat, level = 1)
}
\arguments{
\item{dat}{sam data object}

\item{level}{1 or 2 (1 most basic configuration, 2 configuration with AR correlation structure on surveys)}
}
\value{
a list containing the elements needed to configure a sam model (e.g. minAge, maxAge, maxAgePlusGroup, keyLogFsta, ...).
}
\description{
Setup basic minimal configuration for sam assessment
}
\details{
The configuration returned by defcon is intended as a help to set up a syntactically correct configuration for the sam model. The dimensions are set from the data (years, age-classes, and fleet types available). The configuration is intended to be fairly simplistic in the hope that the model configured will at least converge (not guaranteed). Most importantly: No model validation has been performed, so it should not be assumed that the returned model configuration will result in a sensible assessment of the stock. The actual model configuration is the responsibility of the user.
}
