% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{defpar}
\alias{defpar}
\title{Setup initial values for all model parameters and random effects.}
\usage{
defpar(dat, conf)
}
\arguments{
\item{dat}{sam data object as returned from the function setup.sam.data}

\item{conf}{sam configuration list, which could be read from a configuration file via the \code{loadConf} command, or a default/dummy configuration can be generated via the \code{defcon} function}
}
\value{
a list containing initial values for all model parameters and random effects in the model.
}
\description{
Setup initial values for all model parameters and random effects.
}
\details{
The model parameters and random effects are not initialized in any clever way most are simply set to zero. If convergence problems occour different initial values can be tested, but it is more likely the problem is the model configuration.
}
