% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_referencepoints.R
\name{deterministicReferencepoints}
\alias{deterministicReferencepoints}
\alias{deterministicReferencepoints.sam}
\title{Function to calculate reference points for the embedded deterministic model of a SAM fit}
\usage{
deterministicReferencepoints(fit, referencepoints, ...)

\method{deterministicReferencepoints}{sam}(
  fit,
  referencepoints,
  catchType = "catch",
  nYears = 100,
  Fsequence = seq(0, 2, len = 50),
  aveYears = max(fit$data$years) + (-9:0),
  selYears = max(fit$data$years),
  biasCorrect = FALSE,
  newton.control = list(),
  run = TRUE,
  equilibriumMethod = c("AD", "EC"),
  nosim_ci = 200,
  ncores = 1,
  addSequence = FALSE,
  ...
)
}
\arguments{
\item{fit}{A fitted SAM model}

\item{referencepoints}{list of reference points to calculate (See details)}

\item{...}{other arguments not used}

\item{catchType}{Type of yield to optimize: landing, catch, or discard}

\item{nYears}{Number of years in per-recruit calculations}

\item{Fsequence}{Sequence of F values for plotting and starting values}

\item{aveYears}{Years to average over for biological input}

\item{selYears}{Years to average over for selectivity}

\item{biasCorrect}{Should bias correction be used in \link[TMB]{sdreport}?}

\item{newton.control}{Control arguments passed to the newton optimizer (See \link[TMB]{newton})}

\item{run}{Run estimation? If false, a list of arguments to MakeADFun is returned.}

\item{equilibriumMethod}{Method to use to find equilibrium}

\item{nosim_ci}{Number of simulations for simulation based confidence intervals (only when equilibriumMethod is EC)}

\item{ncores}{Number of cores for simulation}
}
\value{
List of estimated reference points

List of estimated reference points
}
\description{
The function estimates reference points based on deterministic per-recruit calculations with no process variance.
The following reference points are implemented:
\describe{
   \item{F=x}{F fixed to x, e.g., \code{"F=0.3"}}
   \item{StatusQuo}{F in the last year of the assessment}
   \item{StatusQuo-y}{F in the y years before the last in the assessment, e.g., \code{"StatusQuo-1"}}
   \item{MSY}{F that maximizes yield}
   \item{0.xMSY}{Fs that gives 0.x*100\% of MSY, e.g., \code{"0.95MSY"}}
   \item{Max}{F that maximizes yield per recruit}
   \item{0.xdYPR}{F such that the derivative of yield per recruit is 0.x times the derivative at F=0, e.g., \code{"0.1dYPR"}}
   \item{0.xSPR}{F such that spawners per recruit is 0.x times spawners per recruit at F=0, e.g., \code{"0.35SPR"}}
   \item{0.xB0}{F such that biomass is 0.x times the biomass at F=0, e.g., \code{"0.2B0"}}
}
}
\examples{
\dontrun{
  deterministicReferencepoints(fit, c("MSY","0.95MSY","Max","0.35SPR","0.1dYPR","StatusQuo-3"))
}
}
