% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_referencepoints.R
\name{deterministicReferencepoints.sam}
\alias{deterministicReferencepoints.sam}
\title{Function to calculate reference points for the embedded deterministic model of a SAM fit}
\usage{
\method{deterministicReferencepoints}{sam}(
  fit,
  referencepoints,
  catchType = "catch",
  nYears = 300,
  Fsequence = seq(0, 2, len = 50),
  aveYears = max(fit$data$years) + (-9:0),
  selYears = max(fit$data$years),
  biasCorrect = FALSE,
  newton.control = list(),
  run = TRUE,
  ...
)
}
\arguments{
\item{fit}{A fitted SAM model}

\item{referencepoints}{list of reference points to calculate (See details)}

\item{catchType}{landing, catch, or discard}

\item{nYears}{Number of years in per-recruit calculations}

\item{Fsequence}{Sequence of F values for plotting and starting values}

\item{aveYears}{Years to average over for biological input}

\item{selYears}{Years to average over for selectivity}

\item{biasCorrect}{Should bias correction be used in \link[TMB]{sdreport}?}

\item{newton.control}{Control arguments passed to the newton optimizer (See \link[TMB]{newton})}

\item{...}{other arguments not used}
}
\value{

}
\description{
Function to calculate reference points for the embedded deterministic model of a SAM fit
}
\author{
Christoffer Moesgaard Albertsen
}
