% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelforecast.R
\name{.SAM_replicate}
\alias{.SAM_replicate}
\title{Parallel replicate for modelforecast}
\usage{
.SAM_replicate(
  n,
  expr,
  simplify = "array",
  ncores = 1,
  env = parent.frame(n + 1),
  par_precall = NULL,
  type = ifelse(.Platform$OS.type == "unix", "mclapply", "PSOCK")
)
}
\arguments{
\item{n}{number of replicates}

\item{expr}{expression}

\item{simplify}{simplify passes to sapply}

\item{ncores}{number of cores}

\item{env}{environment}

\item{par_precall}{Code to run when starting}

\item{type}{type of parallelisation}
}
\value{
output
}
\description{
Parallel replicate for modelforecast
}
