% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{ellipse}
\alias{ellipse}
\title{A generic function returning an ellipse or other outline of a
confidence region for two parameters.}
\usage{
ellipse(x, scale = c(1, 1), centre = c(0, 0), level = 0.95,
  t = sqrt(qchisq(level, 2)), which = c(1, 2), npoints = 100, ...)
}
\arguments{
\item{x}{An object. In the default method the parameter 'x' should be
a correlation between -1 and 1 or a square positive definite
matrix at least 2x2 in size. It will be treated as the
correlation or covariance of a multivariate normal
distribution.}

\item{scale}{If 'x' is a correlation matrix, then the standard deviations
of each parameter can be given in the scale parameter.  This
defaults to 'c(1, 1)', so no rescaling will be done.}

\item{centre}{The centre of the ellipse will be at this position.}

\item{level}{The confidence level of a pairwise confidence region.  The
default is 0.95, for a 95% region.  This is used to control
the size of the ellipse being plotted.  A vector of levels
may be used.}

\item{t}{The size of the ellipse may also be controlled by specifying
the value of a t-statistic on its boundary.  This defaults to
the appropriate value for the confidence region.}

\item{which}{This parameter selects which pair of variables from the
matrix will be plotted.  The default is the first 2.}

\item{npoints}{The number of points used in the ellipse.  Default is 100.}

\item{...}{Descendant methods may require additional parameters.}
}
\value{
An 'npoints' x '2' matrix is returned with columns named according
to the row names of the matrix 'x' (default ''x'' and ''y''),
suitable for plotting.
}
\description{
Taken from the 'ellipse' package.
}

