% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{empirobscorrplot}
\alias{empirobscorrplot}
\title{Plots the residual between-age correlation matrices by fleet.}
\usage{
empirobscorrplot(res, ...)
}
\arguments{
\item{res}{the object returned from residuals.sam}

\item{...}{extra arguments to plot}
}
\description{
Plots the residual between-age correlation matrices by fleet.
}
