% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{equilibriumbiomassplot}
\alias{equilibriumbiomassplot}
\alias{equilibriumbiomassplot.default}
\alias{equilibriumbiomassplot.samforecast}
\alias{equilibriumbiomassplot.hcr}
\title{SAM equilibrium biomass plot}
\usage{
equilibriumbiomassplot(fit, ...)

\method{equilibriumbiomassplot}{default}(fit, ...)

\method{equilibriumbiomassplot}{samforecast}(fit, ...)

\method{equilibriumbiomassplot}{hcr}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM equilibrium biomass plot
}
\details{
Plot of deterministic equilibrium spawners per recruit assuming biological parameters and selectivity for that year remains unchanged in the future.
}
