% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{equilibriumbiomasstable}
\alias{equilibriumbiomasstable}
\alias{equilibriumbiomasstable.default}
\title{equilibrium biomass table}
\usage{
equilibriumbiomasstable(fit, ...)

\method{equilibriumbiomasstable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
equilibrium biomass table
}
\details{
...
}
