% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{erbplot}
\alias{erbplot}
\alias{erbplot.default}
\alias{erbplot.samforecast}
\alias{erbplot.hcr}
\title{SAM effective reproductive output biomass plot}
\usage{
erbplot(fit, ...)

\method{erbplot}{default}(fit, ...)

\method{erbplot}{samforecast}(fit, ...)

\method{erbplot}{hcr}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM effective reproductive output biomass plot
}
\details{
Plot of spawning stock biomass
}
