% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{erbtable}
\alias{erbtable}
\alias{erbtable.default}
\title{Effective reproductive biomass table}
\usage{
erbtable(fit, ...)

\method{erbtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
Effective reproductive biomass table
}
\details{
...
}
