% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fbartable}
\alias{fbartable}
\alias{fbartable.default}
\title{Fbar table}
\usage{
fbartable(fit, ...)

\method{fbartable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
Fbar table
}
\details{
...
}
