% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast2_sf.R
\name{forecast2}
\alias{forecast2}
\title{Forecast function with annual management strategies}
\usage{
forecast2(fit, fscale = NULL, catchval = NULL, catchval.exact = NULL,
  fval = NULL, MSYBtrig = NULL, Blim = NULL, Fmsy = NULL,
  Fscenario = NULL, Flow = NULL, nextssb = NULL, landval = NULL,
  cwF = NULL, nosim = 1000, year.base = max(fit$data$years),
  ave.years = max(fit$data$years) + (-4:0),
  rec.years = max(fit$data$years) + (-9:0), label = NULL,
  overwriteSelYears = NULL, deterministic = FALSE,
  customWeights = NULL, customSel = NULL, lagR = FALSE,
  splitLD = FALSE, addTSB = FALSE, RW = FALSE, Rdist = FALSE,
  F.RW = TRUE)
}
\arguments{
\item{fit}{an assessment object of type sam, as returned from the function sam.fit}

\item{fscale}{a vector of f-scales. See details.}

\item{catchval}{a vector of target catches. See details.}

\item{catchval.exact}{a vector of target catches which will be met without noise. See details.}

\item{fval}{a vector of target f values. See details.}

\item{MSYBtrig}{a vector of MSYBtrigger values to inform assymptote of the HCRs.}

\item{Blim}{a vector of Blim valuesfor the HCRs.}

\item{Fmsy}{a scalar giving the value of FMSY or the Ftarget that corresponds to MSYBtrigger in the HCRs.}

\item{Fscenario}{a scalar between 1 and 5 informing the shape of the HCR. See details.}

\item{Flow}{a scalar giving the minimum F for Fscenario=3, 4 and 5.}

\item{nextssb}{a vector target SSB values the following year. See details}

\item{landval}{a vector of target catches. See details.}

\item{cwF}{a vector target custom weighted F values. customWeights must also be specified}

\item{nosim}{number of simulations default is 1000}

\item{year.base}{starting year default last year in assessment. Currently it is only supported to use last assessment year or the year before}

\item{ave.years}{vector of years to average for weights, maturity, M and such}

\item{rec.years}{vector of years to use to resample recruitment from}

\item{label}{optional label to appear in short table}

\item{overwriteSelYears}{if a vector of years is specified, then the average selectivity of those years is used (not recommended)}

\item{deterministic}{option to turn all process noise off (not recommended, as it will likely cause bias)}

\item{customWeights}{a vector of same length as number of age groups giving custom weights (currently only used for weighted average of F calculation)}

\item{customSel}{supply a custom selection vector that will then be used as fixed selection in all years after the final assessment year (not recommended)}

\item{lagR}{if the second youngest age should be reported as recruits}

\item{splitLD}{if TRUE the result is split in landing and discards}

\item{addTSB}{if TRUE the total stock biomass (TSB) is added}

\item{RW}{TRUE or FALSE statement specifying if random walk recrutiment should be used in the forecast}

\item{Rdist}{TRUE or FALSE statement specifying if recruitment should follow a normal distribution on the log scale with mean the mean of recruitment in rec.years and variance the variance of recruitment in rec.years}

\item{F.RW}{TRUE or FALSE statement specifying if the F-process should be on in the forecast}
}
\value{
an object of type samforecast
}
\description{
Forecast function modified from the original function (forecast()) by adding the options of annual management strategies.

The management strategies include 5 HCRs (Fscenario) and 3 recruitment options (sampled in year interval (default), lognormally distributed given mean of year interval (Rdist) or random walk (RW)). Different Ftargets and reference points can be set up to parameterize the HCRs (see Fmsy, Flow, Blim, MSYBtrig).
}
\details{
There are four ways to specify a scenario. If e.g. four F values are specified (e.g. fval=c(.1,.2,.3,4)), then the first value is used in the last assessment year (base.year), and the three following in the three following years. Alternatively F's can be specified by a scale, or a target catch. Only one option can be used per year. So for instance to set a catch in the first year and an F-scale in the following one would write catchval=c(10000,NA,NA,NA), fscale=c(NA,1,1,1). The length of the vector specifies how many years forward the scenarios run.

The HCRs have 5 shapes depending on the Fscenario. All HCRs follow an assymptote at Fmsy when SSB>= MSYBtrig. All HCRs except Fscenario=5 have F decreasing linearly between MSYBtrig and Blim. 

Fscenario=1 is the MSY approach where F=0 when SSB<Blim. 

Fscenario=2 is a simple hockey-stick shape where F decrease from Fmsy linearly to 0 below MSYBtrig. 

Fscenario=3 is the scenario where F=Flow when SSB<=Blim. 

Fscenario=4 has the same shape than Fscenario=2 except that F cannot decrease below Flow. 

Fscenario=5 is the case where F=Flow when SSB<=Blim but when Blim<SSB<MSYBtrig F decreases from Fmsy to Flow linearly.
}
