% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMSY.R
\name{forecastMSY}
\alias{forecastMSY}
\alias{forecastMSY.sam}
\title{Estimating Fmsy}
\usage{
forecastMSY(fit, nYears = 100, nlminb.control = list(eval.max = 100,
  iter.max = 100), rec.years = c(), processNoiseF = FALSE, ...)

\method{forecastMSY}{sam}(fit, nYears = 100,
  nlminb.control = list(eval.max = 100, iter.max = 100),
  rec.years = c(), processNoiseF = FALSE, ...)
}
\arguments{
\item{fit}{a SAM fit}

\item{nYears}{Number of years to forecast}

\item{nlminb.control}{list of control variables for nlminb}

\item{...}{other arguments}
}
\description{
Work in progress - do not use
}
\author{
Christoffer Moesgaard Albertsen
}
