% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fsbtable}
\alias{fsbtable}
\alias{fsbtable.default}
\title{FSB table}
\usage{
fsbtable(fit, ...)

\method{fsbtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
FSB table
}
\details{
...
}
