% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{generationlengthtable}
\alias{generationlengthtable}
\alias{generationlengthtable.default}
\title{Generation length table}
\usage{
generationlengthtable(fit, ...)

\method{generationlengthtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
Generation length table
}
\details{
...
}
