% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{getAllDerivedValues}
\alias{getAllDerivedValues}
\title{Update sam fit with additional derived values}
\usage{
getAllDerivedValues(fit)
}
\arguments{
\item{fit}{sam fit returned by sam.fit}
}
\value{
Updated sam fit
}
\description{
Update sam fit with additional derived values
}
