% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{getFleet}
\alias{getFleet}
\title{Extract a fleet from a fitted object}
\usage{
getFleet(fit, fleet)
}
\arguments{
\item{fit}{A fitted object as returned from sam.fit}

\item{fleet}{The number of the fleet}
}
\description{
Extract a fleet from a fitted object
}
