% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{getLowerBounds}
\alias{getLowerBounds}
\title{Bounds}
\usage{
getLowerBounds(parameters)
}
\arguments{
\item{parameters}{initial values for the model in a format similar to what is returned from the defpar function}
}
\value{
a named list
}
\description{
Bounds
}
