% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{getResidualFleets}
\alias{getResidualFleets}
\title{Extract a list of catch fleets}
\usage{
getResidualFleets(fit, pred = "FALSE")
}
\arguments{
\item{fit}{A fitted object as returned from sam.fit}

\item{pred}{Should it be predicted value, default is observed}
}
\value{
A list of matrices of observed or predicted values for catch fleets
}
\description{
Extract a list of catch fleets
}
