% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr}
\alias{hcr}
\alias{hcr.sam}
\title{Harvest control rule forecast}
\usage{
hcr(fit, ...)

\method{hcr}{sam}(
  fit,
  nYears = 20,
  Ftarget,
  Btrigger,
  Forigin = 0,
  Borigin = 0,
  Fcap = 0,
  Bcap = 0,
  nosim = 10000,
  ave.years = max(fit$data$years) + (-4:0),
  rec.years = numeric(0),
  preForecast = list(),
  currentSSB = FALSE,
  ...
)
}
\arguments{
\item{fit}{A SAM fit}

\item{...}{additional arguments passed to \link{modelforecast}}

\item{nYears}{Number of years to forecast}

\item{Ftarget}{Target F for high SSB}

\item{Btrigger}{SSB that triggers the control rule}

\item{Forigin}{F used for SSB = Borigin}

\item{Borigin}{Between Blim and Btrigger, F values are selected based on linear interpolation from Forigin to Ftarget}

\item{Fcap}{F for SSB < Bcap}

\item{Bcap}{SSB for which Fcap is used below}

\item{nosim}{Number of simulations to do. If NULL a model forecast based on the Laplace approximation is used}

\item{ave.years}{vector of years to average for weights, maturity, M and such}

\item{rec.years}{vector of years to use to resample recruitment from. If an empty vector is given, recruitment is based on the fitted model.}

\item{preForecast}{list of forecast parameters (i.e., fval, fscale, catchval, landval, or nextssb) to use before the HCR}

\item{currentSSB}{if TRUE, SSB at the begining of the control rule year is used. If FALSE, SSB at the begining of the previous year is used.}
}
\value{
model forecast using a harvest control rule

hcr model forecast object
}
\description{
The formula below is used to determine a new F based on the previous SSB.
\deqn{F = \left\{
              \begin{array}{ll}
                  F_{cap} & SSB < B_{cap} \\
                  min\left(F_{target}, \max\left( F_{origin}, (SSB - B_{origin}) \cdot (F_{target} - F_{origin}) / (B_{trigger}-B_{origin}) \right)\right) & SSB \ge B_{cap}
              \end{array}\right.
}
If \eqn{B_{trigger} = B_{origin}} and \eqn{SSB \ge B_{cap}}, \eqn{F_{target}} is always returned.
}
\seealso{
modelforecast
}
