% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{ICESvalues}
\alias{ICESvalues}
\alias{ICESvalues.sam}
\alias{ICESvalues.sam_referencepoints}
\title{Function to estimate ICES values}
\usage{
ICESvalues(x, nosim, nyears, ntail, ...)

\method{ICESvalues}{sam}(x, nosim, nyears, ...)

\method{ICESvalues}{sam_referencepoints}(
  x,
  nosim = 1000,
  nyears = 100,
  ntail = 10,
  quantile_buffer_Fmsy = 0.5,
  calculate_Fp05 = TRUE,
  ...
)
}
\arguments{
\item{x}{object to calculate for}

\item{nosim}{Number of simulations}

\item{nyears}{Number of years to forecast before equilibrium}

\item{ntail}{Number of years to use for calculations}

\item{...}{Other parameters passed to \link{referencepoints}}
}
\value{
ICES values
}
\description{
The function calculates MSYBtrigger and Bpa from estimated reference points through a simulation forecast. If a sam object is used, other reference points (e.g., Fmsy) are estimated first.
}
\author{
Christoffer Moesgaard Albertsen
}
