% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineKnots.R
\name{iibc}
\alias{iibc}
\title{Double integrated spline basis for use with formula interface}
\usage{
iibc(x, df = 3L, knots = NULL, Boundary.knots = range(x), intercept = FALSE)
}
\arguments{
\item{x}{Points to evaluate the basis in}

\item{df}{Degrees of freedom}

\item{knots}{Internal knots. If NULL, they are selected from quantiles of x.}

\item{Boundary.knots}{Boundary knots. Defaults to range of x}

\item{intercept}{Include an intercept in basis?}
}
\value{
A spline basis
}
\description{
Double integrated spline basis for use with formula interface
}
