% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{is.whole.positive.number}
\alias{is.whole.positive.number}
\title{Function to test if x is ...}
\usage{
is.whole.positive.number(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{number}

\item{tol}{precision}
}
\description{
Function to test if x is ...
}
\details{
...
}

