% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{jit}
\alias{jit}
\alias{jit.sam}
\title{Jitter runs}
\usage{
jit(fit, nojit = 10, ...)

\method{jit}{sam}(
  fit,
  nojit = 10,
  par = defpar(fit$data, fit$conf),
  sd = 0.25,
  ncores = detectCores(),
  ...
)
}
\arguments{
\item{fit}{a fitted model object as returned from sam.fit}

\item{nojit}{a list of vectors. Each element in the list specifies a run where the fleets mentioned are omitted}

\item{par}{initial values to jitter around. The defaule ones are returned from the defpar function}

\item{sd}{the standard deviation used to jitter the initial values (most parameters are on a log scale, so similar to cv)}

\item{ncores}{the number of cores to attemp to use}
}
\value{
A "samset" object, which is basically a list of sam fits
}
\description{
Jitter runs
}
\details{
...
}
