% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lifeexpectancyplot}
\alias{lifeexpectancyplot}
\alias{lifeexpectancyplot.default}
\alias{lifeexpectancyplot.samforecast}
\alias{lifeexpectancyplot.hcr}
\title{SAM life expectancy plot}
\usage{
lifeexpectancyplot(fit, ...)

\method{lifeexpectancyplot}{default}(fit, ...)

\method{lifeexpectancyplot}{samforecast}(fit, ...)

\method{lifeexpectancyplot}{hcr}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM life expectancy plot
}
\details{
Plot of life expectancy
}
