% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{lifeexpectancytable}
\alias{lifeexpectancytable}
\alias{lifeexpectancytable.default}
\title{SSB table}
\usage{
lifeexpectancytable(fit, ...)

\method{lifeexpectancytable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
SSB table
}
\details{
...
}
