% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{loadConf}
\alias{loadConf}
\title{Loads a model configuration from a file}
\usage{
loadConf(dat, file, patch = FALSE)
}
\arguments{
\item{dat}{sam data list as returned from the function setup.sam.data}

\item{file}{the file to read the configuration from}

\item{patch}{logical if TRUE missing entries will be automatically filled with default values}
}
\description{
Loads a model configuration from a file
}
\details{
function useful loading a model configuration. Such a configuration can be saved via the saveConf function
}
