% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.sam}
\alias{logLik.sam}
\title{Log likelihood of sam object}
\usage{
\method{logLik}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object as returned from the \code{\link{sam.fit}} function}

\item{...}{extra arguments}
}
\description{
Log likelihood of sam object
}
\details{
log likelihood of a sam model run
}
