% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{modelDescription}
\alias{modelDescription}
\title{Description of model}
\usage{
modelDescription(fit, ...)
}
\arguments{
\item{fit}{returned object from sam.fit}

\item{...}{Additional parameters to be passed to ...}
}
\description{
Description of model
}
\details{
...
}
