% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelforecast.R
\name{modelforecast}
\alias{modelforecast}
\alias{modelforecast.sam}
\title{Model based forecast function}
\usage{
modelforecast(fit, ...)

\method{modelforecast}{sam}(
  fit,
  constraints = NULL,
  fscale = NULL,
  catchval = NULL,
  fval = NULL,
  nextssb = NULL,
  landval = NULL,
  nosim = 0,
  year.base = max(fit$data$years),
  ave.years = max(fit$data$years) + (-9:0),
  rec.years = c(),
  rec.fixval = numeric(0),
  rec.fixvalVar = (1e-06)^2,
  label = NULL,
  overwriteSelYears = NULL,
  deterministicF = FALSE,
  processNoiseF = FALSE,
  fixedFdeviation = FALSE,
  useFHessian = FALSE,
  resampleFirst = !is.null(nosim) && nosim > 0,
  useModelLastN = TRUE,
  customSel = NULL,
  lagR = FALSE,
  splitLD = FALSE,
  addTSB = FALSE,
  biasCorrect = FALSE,
  returnAllYears = FALSE,
  returnObj = FALSE,
  progress = TRUE,
  estimate = median,
  silent = TRUE,
  newton_config = NULL,
  custom_pl = NULL,
  useNonLinearityCorrection = (nosim > 0 && !deterministicF),
  ncores = 1,
  mc.type = "mclapply",
  useRecPool = FALSE,
  overwriteBioProcessModel = FALSE,
  useManagementLag = FALSE,
  assessmentErrorMean_F = 0,
  assessmentErrorRho_F = 0,
  assessmentErrorSigma_F = 0,
  assessmentErrorMean_N = 0,
  assessmentErrorRho_N = 0,
  assessmentErrorSigma_N = 0,
  assessmentErrorMean_M = 0,
  assessmentErrorRho_M = 0,
  assessmentErrorSigma_M = 0,
  assessmentErrorMean_Mat = 0,
  assessmentErrorRho_Mat = 0,
  assessmentErrorSigma_Mat = 0,
  assessmentErrorMean_SW = 0,
  assessmentErrorRho_SW = 0,
  assessmentErrorSigma_SW = 0,
  assessmentErrorMean_CW = 0,
  assessmentErrorRho_CW = 0,
  assessmentErrorSigma_CW = 0,
  implementationErrorRho_F = 0,
  ...
)
}
\arguments{
\item{fit}{SAM model fit}

\item{...}{other variables used by the methods}

\item{constraints}{a character vector of forecast constraint specifications}

\item{fscale}{a vector of f-scales. See details.}

\item{catchval}{a vector of target catches. See details "old specification".}

\item{fval}{a vector of target f values. See details "old specification".}

\item{nextssb}{a vector target SSB values the following year. See details "old specification".}

\item{landval}{a vector of target catches. See details "old specification".}

\item{nosim}{number of simulations. If 0, the Laplace approximation is used for forecasting.}

\item{year.base}{starting year default last year in assessment. Currently it is only supported to use last assessment year or the year before}

\item{ave.years}{vector of years to average for weights, maturity, M and such}

\item{rec.years}{vector of years to use to resample recruitment from. If the vector is empty, the stock recruitment model is used.}

\item{label}{optional label to appear in short table}

\item{overwriteSelYears}{if a vector of years is specified, then the average selectivity of those years is used (not recommended)}

\item{deterministicF}{option to set F variance to (almost) zero (not recommended)}

\item{processNoiseF}{option to turn off process noise in F}

\item{fixedFdeviation}{Use a fixed F deviation from target?}

\item{useFHessian}{Use the covariance of F estimates instead of the estimated process covariance for forecasting?}

\item{resampleFirst}{Resample base year when nosim > 0?}

\item{useModelLastN}{Use last N?}

\item{customSel}{supply a custom selection vector that will then be used as fixed selection in all years after the final assessment year (not recommended)}

\item{lagR}{if the second youngest age should be reported as recruits}

\item{splitLD}{if TRUE the result is split in landing and discards}

\item{addTSB}{if TRUE the total stock biomass (TSB) is added}

\item{biasCorrect}{Do bias correction of reported variables. Can be turned off to reduce running time (not recommended).}

\item{returnAllYears}{If TRUE, all years are bias corrected. Otherwise, only forecast years are corrected.}

\item{returnObj}{Only return TMB object?}

\item{progress}{Show progress bar for simulations?}

\item{estimate}{the summary function used (typically mean or median) for simulations}

\item{silent}{Passed to MakeADFun. Should the TMB object be silent?}

\item{newton_config}{Configuration for newton optimizer to find F values. See ?TMB::newton for details. Use NULL for TMB defaults.}

\item{custom_pl}{Parameter list. By default, the parameter list from fit is used.}

\item{useNonLinearityCorrection}{Should a non linearity correction be added to transformation of logF? See Details - Non-linearity correction.}

\item{ncores}{Number of cores to use if simulating}
}
\value{
an object of type samforecast
}
\description{
Model based forecast function

Model based forecast function
}
\details{
Function to forecast the model under specified catch constraints. In the forecast, catch constraints are used to set the mean of the \eqn{log(F)} process for each simulation. Therefore, catch constraints are not matched exactly in individual simulations. Likewise, the summary of a specific set of simulations will not match exactly due to random variability.
By default, recruitment is forecasted using the estimated recruitment model. If a vector of recruitment years is given, recruitment is forecasted using a log-normal distribution with the same mean and variance as the recruitment in the years given. This is different from the forecast function, which samples from the recruitment estimates.
Catch scenarios are specified by a vector of target constraints. The first value determines F in the year after the base year.
}
\section{Forecast constraints}{


\subsection{F based constraints:}{
Forecasts for F values are specified by the format "F[f,a0-a1]=x" where f is the residual catch fleet and a0-a1 is an age range. For example, "F[2,2-4]=0.3" specifies that the average F for the second fleet over ages 2-4 should be 0.3.
If an "*" is added to the target value, the target will be relative to the year before. For example, "F[2,2-4]=0.9*" specifies that the average F for the second fleet over ages 2-4 should be 90% of the year before. Further, the target for a fleet can be relative to the total by adding "*F" or to another fleet by adding "*F[f]" where f is the fleet number. The same age range will always be used.
If the fleet is omitted (e.g., F[2-4]), the target is for the total F.
If the age range is omitted (e.g., F[2]), the fbar range of the model is used.
Likewise, both fleet and age range can be omited (e.g., F=0.3) to specify a value for total F with the range used in the model.

For example:
\describe{
\item{"F=0.2"}{Will set the median average total fishing mortality rate to 0.2}
\item{"F[1]=0.2"}{Will set the median average fishing mortality rate of the first fleet to 0.2}
\item{"F[2-4]=0.2"}{Will set the median average total fishing mortality rate over ages 2 to 4 to 0.2}
\item{"F[3,2-4]=0.2"}{Will set the median average fishing mortality rate over ages 2 to 4 for the third fleet to 0.2}
}
}

\subsection{Catch/Landing based constraints:}{
Forecasts for catch and landing values are specified by the format "C[f,a0-a1]=x" for catch and "L[f,a0-a1]" for landings. If the age range is omitted, all modelled ages are used. Otherwise, the format is similar to F based scenarios.
If an "*" is added to the target value, the target will be relative to the year before.
Further, the catch target for a fleet can be relative to the total by adding "*C" or to another fleet by adding "*C[f]" where f is the fleet number. The same age range will always be used. Likewise, relative landing targets can be specified using "*", "*L", or "*L[f]" for targets relative to last year, the total, or fleet f, respectively.

For example:
\describe{
\item{"C=100000"}{Will scale F such that the total predicted catch is 100000}
\item{"C[1]=100000"}{Will scale F such that the predicted catch of the first fleet is 100000}
\item{"C[2-4]=100000"}{Will scale F such that the total predicted catch for ages 2 to 4 is 100000}
\item{"C[3,2-4]=100000"}{Will scale F such that the predicted catch for ages 2 to 4 in the third fleet is 100000}
\item{"L=100000"}{Will scale F such that the total predicted landing is 100000}
\item{"L[1]=100000"}{Will scale F such that the predicted landing of the first fleet is 100000}
\item{"L[2-4]=100000"}{Will scale F such that the total predicted landing for ages 2 to 4 is 100000}
\item{"L[3,2-4]=100000"}{Will scale F such that the predicted landing for ages 2 to 4 in the third fleet is 100000}
}
}

\subsection{Next year's SSB/TSB based constraints:}{
Forecasts for spawning stock biomass (SSB) and total stock biomass (TSB) values are specified by the format "SSB[a0-a1]=x" for SSB and "TSB[a0-a1]" for TSB. For setting F in year y, the relevant biomass for year y+1 is predicted for the constraint. If spawning is not at the beginning of the year, F is assumed to be the same for year y and y+1 in the prediction.
The format is similar to catch/landing based scenarios. However, fleets have no effect. If an age range is omitted, the full age range of the model is used.
If an "*" is added to the target value, the target will be relative to the year before. That is, when setting F in year y, the predicted biomass in year y+1 will be relative to the biomass in year y-1.
Note that since SSB and TSB used for catch constraints are predicted, the input constraint will differ from the output SSB and TSB estimates due to process variability.

For example:
\describe{
\item{SSB=200000}{Will scale F such that the predicted SSB at the beginning of the next year is 200000}
\item{SSB[3-9]=200000}{Will scale F such that the predicted SSB for ages 3 to 9 at the beginning of the next year is 200000}
\item{TSB=200000}{Will scale F such that the predicted TSB at the beginning of the next year is 200000}
\item{TSB[3-9]=200000}{Will scale F such that the predicted TSB for ages 3 to 9 at the beginning of the next year is 200000}
}
}

\subsection{Harvest control rule based constraints:}{
Harvest control rules can be specified for forecasts using the format "HCR=x~y" where x is the target and y is the biomass trigger (see ?hcr for full details on the form of the harvest control rule). Further, the target can be specified as an F target ("HCR=xF~y"), catch target ("HCR=xC~y"), or landing target ("HCR=xL~y"). Likewise the trigger can either be for SSB ("HCR=x~ySSB") or TSB ("HCR=x~yTSB"). Age ranges can be set for both triggers and targets and a fleet can be set for the target. The notation and defaults are similar to the F based and SSB/TSB based constraints, respectively.
When setting F in year y, the projected biomass in year y is used by default. To use the (at this time known) biomass in a previous year, a time lag can be specified. To specify a time lag of, e.g., 1 year for SSB the format is "HCR=x~ySSB-1".
Finally, the origin and cap for the HCR can be set using "HCR[FO=a,FC=b,BO=d,BC=e]=x~y", where FO is the F (or catch or landing) value at origin, BO is the biomass at origin, FC is the F (or catch or landing) value when the HCR is capped and BC is the biomass at which the HCR is capped. See ?hcr for further details on the shape of the HCR.
For a HCR similar to the ICES advice rule, the specification is on the form "HCR[BC=Blim] = fmsy~MSYBtrigger". Note that, unlike an ICES advice rule, the HCR does not do a forecast to determine if fishing can continue below Blim.

For example:
\describe{
\item{HCR=0.9~100000}{Will apply a harvest control rule with an F target of 0.9 and a biomass trigger of 100000 on SSB}
\item{HCR=10000C~100000}{Will apply a harvest control rule with a catch target of 10000 and a biomass trigger of 100000 on SSB}
\item{HCR=0.9~100000SSB}{Will apply a harvest control rule with an F target of 0.9 and a biomass trigger of 100000 on SSB}
\item{HCR=0.9F[1,2-4]~100000SSB}{Will apply a harvest control rule with an F target on the first fleet ages 2-4 of 0.9 and a biomass trigger of 100000 on SSB}
\item{HCR=0.9~100000TSB[0-4]}{Will apply a harvest control rule with an F target of 0.9 and a biomass trigger of 100000 on TSB for ages 0 to 4}
\item{HCR[FC=1e-9,BC=20000]=0.9~100000}{Will apply a harvest control rule with an F target of 0.9 and a biomass trigger of 100000 on SSB where biomass values below 20000 will give an F of 1e-9}
\item{HCR[FO=0,BO=30000]=0.9~100000}{Will apply a harvest control rule with an F target of 0.9 and a biomass trigger of 100000 on SSB where the slope on which F is reduced goes to zero F at a biomass of 30000}
}
}

\subsection{Combining constraints:}{
Constraints for different fleets can be combined by "&".
For example, "F[2-4]=0.5 & C[2]=10000" specifies that total Fbar over ages 2-4 should be 0.5 while the catch for the second residual catch fleet should be 10,000t.
The constraints cannot affect within-fleet selectivity. Therefore, a fleet can at most have one constraint per year, and the total number of constraints cannot exceed the number of catch fleets. That is, if a constraint is given for the sum of fleets, there must be at least one fleet without any constraints.
For fleets where no constraints are given, a constraint is set to keep their relative Fs constant.
}

\subsection{Values relative to previous year:}{
Catch constraints specified as specific values are inherently different from catch constraints specified as relative values, even if they lead to the same F. Catch constraints specified as relative values will propagate the uncertainty in, e.g, F from previous years whereas constraints specified as specific values will not. This is different from the \link{forecast} function where, for example, a forecast using fval is the same as a forecast using fscale, if they lead to the same F. 
}


##' \subsection{Process variability:}{
In the forecast, constraints are used to set the predicted F value in year y based on information available until year y-1. Therefore, constraints using predicted values for year y, such as catch, will not be matched exactly by the realized catch due to process variability in F, N, biological processes and catch itself.
}
}

\section{Non-linearity correction}{

In the model forecasts, constraints are calculated to set the mean of the log(F) process, corresponding to the median F-at-ages. Typically, the constraints are non-linear functions of log(F)-at-age. Therefore, when stochasticity is added to log(F) (i.e., deterministicF=FALSE), target values will correspond to a transformation of the median, and not the median of the transformation. For example, a target for the average fishing mortality (Fbar) will correspond to the average of the median F at age, which will be different from the median Fbar.

The "useNonLinearityCorrection" argument can be used to shift the target from a function of the mean log(F) (median F) towards the log-mean of the function of log(F), which is approximately the median of the function of log(F).
}

\section{Old specification}{

It is also possible to specify forecast constraints in a way similar to the \link{forecast} function. 
There are four ways to specify a scenario. If e.g. four F values are specified (e.g. fval=c(.1,.2,.3,4)), then the first value is used in the year after the last assessment year (base.year + 1), and the three following in the three following years. Alternatively F's can be specified by a scale, or a target catch. Only one option can be used per year. So for instance to set a catch in the first year and an F-scale in the following one would write catchval=c(10000,NA,NA,NA), fscale=c(NA,1,1,1). If only NA's are specified in a year, the F model is used for forecasting. The length of the vector specifies how many years forward the scenarios run. Unlike the forecast function, no value should be given for the base year.
Internally, the old specification is translated such that "fval=x" becomes "F=x", "fscale=x" becomes "F=x*", "catchval=x" becomes "C=x", "nextssb=x" becomes "SSB=x", and "landval=x" becomes "L=x".
}

\section{Forecasts using Laplace approximation or simulations}{

Forecasts can be made using either a Laplace approximation projection (nosim=0) or simulations (nosim > 0). When using the Laplace approximation, the most likely projected trajectory of the processes along with a confidence interval is returned. In contrast, simulation based forecasts will return individual simulated trajectories and summarize using the function given as the estimate argument along with an interval covering 95% of the simulations.
}

\section{Warnings}{

Long term forecasts with random walk recruitment can lead to unstable behaviour and difficulties finding suitable F values for the constraints.
If no suitable F value can be found, an error message will be shown, and F values will be NA or NaN. Likewise, forecasts leading to high F values in some years (or large changes from one year to another) may cause problems for the optimization as they will be used as starting values for the next years.
Since the model works on log space, all target values should be strictly positive. Values too close to zero may cause problems.
}

\seealso{
forecast
}
