% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{modelVersionInfo}
\alias{modelVersionInfo}
\title{Description of model}
\usage{
modelVersionInfo(fit, ...)
}
\arguments{
\item{fit}{returned object from sam.fit}

\item{...}{Additional parameters to be passed to ...}
}
\description{
Description of model
}
\details{
Writes a string to install the version of the package which was used to run the model.
}
