% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{mohn}
\alias{mohn}
\title{Mohn's rho calculation}
\usage{
mohn(fits, what = NULL, lag = 0, ...)
}
\arguments{
\item{fits}{a samset object as returned from the retro function.}

\item{what}{a function computing the quantity to calculate Mohn's rho for (default NULL computes Fbar, SSB, and R).}

\item{lag}{lag applied to fits and reference fit.}

\item{...}{extra arguments}
}
\description{
Mohn's rho calculation
}
\details{
...
}
