% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{MSE}
\alias{MSE}
\title{Management strategy evaluation using SAM models}
\usage{
MSE(
  OM,
  EM,
  nYears,
  AdviceForecastSettings,
  AdviceYears = 1,
  AdviceLag = 0,
  initialAdvice = NA,
  implementationError = function(x) x,
  knotRange = 3,
  intermediateFleets = numeric(0),
  OMselectivityFixed = FALSE,
  ...
)
}
\arguments{
\item{OM}{sam.fit that will work as operating model}

\item{EM}{sam.fit that will work as estimation model}

\item{nYears}{Number of years to run simulation}

\item{AdviceYears}{Number of years advice given at a time. How advice is given is determined by forecastSettings}

\item{AdviceLag}{Lag between assessment and advice}

\item{initialAdvice}{Advice in the first AdviceLag years}

\item{implementationError}{Function to add implementation error (i.e, transform advice to target catch)}

\item{knotRange}{Range of spline knot values to try}

\item{intermediateFleets}{Fleets that are available in the (first) intermediate year}

\item{...}{arguments passed on to addSimulatedYears}

\item{forecastSettings}{Settings to do forecast that determines advice}
}
\value{
a list with MSE result
}
\description{
Management strategy evaluation using SAM models
}
