% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.sam}
\alias{nobs.sam}
\title{Extract number of observations from sam object}
\usage{
\method{nobs}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object (result from sam.fit)}

\item{...}{extra arguments}
}
\description{
Extract number of observations from sam object
}
\details{
...
}

