\name{nscodConf}
\alias{nscodConf}
\docType{data}
\title{
nscodConf
}
\description{
nscodConf
}
\usage{data("nscodConf")}
\format{
 The format is:
 $ minAge                   
 $ maxAge                   
 $ maxAgePlusGroup          
 $ keyLogFsta               
 $ corFlag                  
 $ keyLogFpar               
 $ keyQpow                  
 $ keyVarF                  
 $ keyVarLogN               
 $ keyVarObs                
 $ stockRecruitmentModelCode
 $ noScaledYears            
 $ keyScaledYears           
 $ keyParScaledYA           
 $ fbarRange                
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(nscodConf)
## maybe str(nscodConf) ; plot(nscodConf) ...
}
\keyword{datasets}
