\name{nscodData}
\alias{nscodData}
\docType{data}
\title{
nscodData
}
\description{
nscodData
}
\usage{data("nscodData")}
\format{
  The format is:
 $ noFleets    
 $ fleetTypes  
 $ sampleTimes 
 $ noYears     
 $ years       
 $ nobs        
 $ idx1        
 $ idx2        
 $ aux
 $ logobs         
 $ propMat   
 $ stockMeanWeight          
 $ catchMeanWeight          
 $ natMor                   
 $ landFrac                 
 $ disMeanWeight            
 $ landMeanWeight           
 $ propF                    
 $ propM                    
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(nscodData)
## maybe str(nscodData) ; plot(nscodData) ...
}
\keyword{datasets}
