\name{nscodParameters}
\alias{nscodParameters}
\docType{data}
\title{
nscodParameters
}
\description{
nscodParameters
}
\usage{data("nscodParameters")}
\format{
  The format is:
List of 14
 $ logFpar     
 $ logQpow     
 $ logSdLogFsta
 $ logSdLogN   
 $ logSdLogObs 
 $ rec_loga    
 $ rec_logb    
 $ itrans_rho   
 $ logScale    
 $ logScaleSSB 
 $ logPowSSB   
 $ logSdSSB    
 $ logF        
 $ logN        
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(nscodParameters)
## maybe str(nscodParameters) ; plot(nscodParameters) ...
}
\keyword{datasets}
