% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{parplot}
\alias{parplot}
\title{SAM parameter plot}
\usage{
parplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM parameter plot
}
\details{
...
}

