% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{partable}
\alias{partable}
\alias{partable.sam}
\title{parameter table}
\usage{
partable(fit, ...)

\method{partable}{sam}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
parameter table
}
\details{
...
}
