% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.hcr}
\alias{plot.hcr}
\title{Plot hcr object}
\usage{
\method{plot}{hcr}(x, ...)
}
\arguments{
\item{x}{output from the \code{\link{hcr}} function}

\item{...}{extra arguments}
}
\description{
Plot hcr object
}
\details{
...
}
