% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.sam}
\alias{plot.sam}
\title{Plot sam object}
\usage{
\method{plot}{sam}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments (not possible to use add=TRUE --- please collect to a list of fits using e.g the c(...), and then plot that collected object)}
}
\description{
Plot sam object
}
\details{
...
}

