% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.samforecast}
\alias{plot.samforecast}
\title{Plot samforecast object}
\usage{
\method{plot}{samforecast}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot samforecast object
}
\details{
...
}
