% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.samset}
\alias{plot.samset}
\title{Plot sam object}
\usage{
\method{plot}{samset}(x, ...)
}
\arguments{
\item{x}{fitted object as returned from the \code{\link{sam.fit}} function.}

\item{...}{extra arguments}
}
\description{
Plot sam object
}
\details{
...
}
