% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotby}
\alias{plotby}
\title{Plot by one or two}
\usage{
plotby(x = NULL, y = NULL, z = NULL, x.line = NULL, y.line = NULL,
  z.line = NULL, by = NULL, bubblescale = 1, x.common = !is.null(x),
  y.common = !is.null(y), z.common = !is.null(z), xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, zmax = NULL, axes = TRUE,
  ...)
}
\arguments{
\item{x}{numeric vector of points to be plotted}

\item{y}{numeric vector of points to be plotted}

\item{z}{numeric vector of points to be plotted}

\item{x.line}{numeric vector of points of line to be added}

\item{y.line}{numeric vector of points of line to be added}

\item{z.line}{numeric vector of points of line to be added}

\item{by}{vector or two column matrix to create sub sets from}

\item{bubblescale}{scaling of bubble size}

\item{x.common}{logical: use same x-axis for all plots}

\item{y.common}{logical: use same y-axis for all plots}

\item{z.common}{logical: use same z-axis for all plots}

\item{xlab}{normal graphical parameter}

\item{ylab}{normal graphical parameter}

\item{xlim}{normal graphical parameter}

\item{ylim}{normal graphical parameter}

\item{zmax}{internally used to scale bubbles similarly}

\item{axes}{normal graphical parameter}

\item{...}{additional graphical parameters}
}
\description{
Plot by one or two
}
\details{
Function used for splitting plots e.g. used to plot residuals
}
\examples{
exdat<-expand.grid(age=1:5, year=1950:2016, fleet=1:3)
exdat$perfectres<-rnorm(nrow(exdat))
attach(exdat)
par(ask=FALSE)
plotby(year,age,perfectres, by=fleet)
detach(exdat)
}
