% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotcorr}
\alias{plotcorr}
\title{This function plots a correlation matrix using ellipse-shaped
glyphs for each entry.  The ellipse represents a level curve of
the density of a bivariate normal with the matching correlation.}
\usage{
plotcorr(corr, outline = TRUE, col = "grey", numbers = FALSE,
  type = c("full", "lower", "upper"), diag = (type == "full"), bty = "n",
  axes = FALSE, xlab = "", ylab = "", asp = 1,
  cex.lab = par("cex.lab"), cex = 0.75 * par("cex"), mar = 0.1 + c(2, 2,
  4, 2), ...)
}
\arguments{
\item{corr}{A matrix containing entries between '-1' and '1' to be plotted as correlations.}

\item{outline}{Whether the ellipses should be outlined in the default colour.}

\item{col}{Which colour(s) to use to fill the ellipses.}

\item{numbers}{Whether to plot numerical correlations in place of ellipses.
If numbers is 'TRUE', then the correlations will be rounded
to a single decimal place and placed on the plot.}

\item{type}{Character. Plot '"full"' matrix or just '"upper"' or
'"lower"' triangular part of it.}

\item{diag}{Logical. Plot diagonal elements or not.}

\item{bty}{for 'plot'}

\item{axes}{for 'plot'}

\item{xlab}{for 'plot'}

\item{ylab}{for 'plot'}

\item{asp}{for 'plot'}

\item{cex.lab}{for 'plot'}

\item{cex}{for 'plot'}

\item{mar}{for 'plot'}

\item{...}{for 'plot'}
}
\description{
Taken from the 'ellipse' package.
}

