% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{predstdplot}
\alias{predstdplot}
\title{Prediction-standard deviation plot}
\usage{
predstdplot(
  fit,
  fleet,
  age = NULL,
  type = "log",
  ylim = NULL,
  ylab = "Standard deviation",
  xlab = "Prediction",
  main = "Pred-std relation",
  ...
)
}
\arguments{
\item{fit}{A sam fit object returned from sam.fit.}

\item{fleet}{Fleet number to plot relation between prediction and standard deviation.}

\item{age}{Relation at age. Only used in cases with more than one relation within the same fleet.}

\item{type}{Either 'log' or 'natural': relation for observations on a log or natural scale.}

\item{ylim}{Optional, sent to plot}

\item{ylab}{Optional, sent to plot}

\item{xlab}{Optional, sent to plot}

\item{main}{Optional, sent to plot}

\item{...}{Sent to plot}
}
\description{
Prediction-standard deviation plot
}
\details{
Plot the relation between observation prediction and standard deviation.
}
