% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.hcr}
\alias{print.hcr}
\title{Print hcr object}
\usage{
\method{print}{hcr}(x, ...)
}
\arguments{
\item{x}{a sam hcr object as returned by \code{\link{hcr}}}

\item{...}{extra arguments}
}
\description{
Print hcr object
}
\details{
prints the HCR forecast
}
