% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.jitset}
\alias{print.jitset}
\title{Print jitset object}
\usage{
\method{print}{jitset}(x, ...)
}
\arguments{
\item{x}{an object as returned from the jit function}

\item{...}{extra arguments}
}
\description{
Print jitset object
}
\details{
...
}
