% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sam}
\alias{print.sam}
\title{Print sam object}
\usage{
\method{print}{sam}(x, ...)
}
\arguments{
\item{x}{the fitted object as returned from the \code{\link{sam.fit}} function}

\item{...}{extra arguments}
}
\description{
Print sam object
}
\details{
prints the log-likelihood and the main convergence criteria
}
