% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sam_referencepoints}
\alias{print.sam_referencepoints}
\title{Print referencepoint object}
\usage{
\method{print}{sam_referencepoints}(x, tables = c("F", "Biomass", "Yield"), digits = 4, format = "f", ...)
}
\arguments{
\item{x}{a sam referencepoint object as returned by \code{\link{referencepoints}}}

\item{tables}{tables to print}

\item{digits}{number of digits to print}

\item{format}{printing format for numbers}

\item{...}{extra arguments}
}
\description{
Print referencepoint object
}
\details{
prints the F reference point table
}
