% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samforecast}
\alias{print.samforecast}
\title{Print samforecast object}
\usage{
\method{print}{samforecast}(x, ...)
}
\arguments{
\item{x}{an object as returned from the forecast function}

\item{...}{extra arguments}
}
\description{
Print samforecast object
}
\details{
...
}
