% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samres}
\alias{print.samres}
\title{Print samres object}
\usage{
\method{print}{samres}(x, ...)
}
\arguments{
\item{x}{a sam residual object as returnd from either \code{\link{residuals.sam}} or \code{\link{procres}}}

\item{...}{extra arguments}
}
\description{
Print samres object
}
\details{
prints the residuals as a data.frame
}
