% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samset}
\alias{print.samset}
\title{Print samset object}
\usage{
\method{print}{samset}(x, ...)
}
\arguments{
\item{x}{a list of sam models}

\item{...}{extra arguments}
}
\description{
Print samset object
}
\details{
...
}
