% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samypr}
\alias{print.samypr}
\title{Print samypr object}
\usage{
\method{print}{samypr}(x, ...)
}
\arguments{
\item{x}{an object as returned from the ypr function}

\item{...}{extra arguments}
}
\description{
Print samypr object
}
\details{
...
}
