% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{qtable}
\alias{qtable}
\title{table of survey catchabilities}
\usage{
qtable(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
table of survey catchabilities
}
\details{
...
}
