% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{qtable.sam}
\alias{qtable.sam}
\title{table of survey catchabilities}
\usage{
\method{qtable}{sam}(fit, ...)
}
\arguments{
\item{fit}{A sam fit as returned from the sam.fit function}

\item{...}{extra arguments not currently used}
}
\description{
table of survey catchabilities
}
