% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{qtableplot}
\alias{qtableplot}
\alias{qtableplot.samqtable}
\title{plot survey catchabilities}
\usage{
qtableplot(qt, exp = FALSE)

\method{qtableplot}{samqtable}(qt, exp = FALSE)
}
\arguments{
\item{qt}{An object of class 'samqtable' as returned from qtable}

\item{exp}{if true return on natural scale rather than log}
}
\description{
plot survey catchabilities

plot survey catchabilities
}
