% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rbtable}
\alias{rbtable}
\alias{rbtable.default}
\title{Reproductive biomass table}
\usage{
rbtable(fit, ...)

\method{rbtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
Reproductive biomass table
}
\details{
...
}
