% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read.data.files}
\alias{read.data.files}
\title{Read all standard data SAM files and return a list as created by 'setup.sam.data'}
\usage{
read.data.files(dir = ".")
}
\arguments{
\item{dir}{Directory to read from}
}
\value{
list (as created by 'setup.sam.data')
}
\description{
Read all standard data SAM files and return a list as created by 'setup.sam.data'
}
\details{
Read all standard SAM data files
}
