% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading.R
\name{read.ices}
\alias{read.ices}
\title{Function to read ICES/CEFAS data files and validate if input makes sense}
\usage{
read.ices(filen)
}
\arguments{
\item{filen}{The filename}
}
\description{
Function to read ICES/CEFAS data files and validate if input makes sense
}
\details{
First two lines are ignored and can be used for comments.
Can read formats 1 full, 2 row, 3 scalar, and 5 column

Tests:
Formatcode is valid, years and ages are pos. integers
minimum <= maximum for years and ages
number of rows and coulmns match year and age ranges
data contains only numbers.

Returns: A validated data matrix.
}

