% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading.R
\name{read.surveys}
\alias{read.surveys}
\title{Function to read ices survey format}
\usage{
read.surveys(filen)
}
\arguments{
\item{filen}{the file}
}
\description{
Function to read ices survey format
}
\details{
...
}

