% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read.table.nowarn}
\alias{read.table.nowarn}
\title{Function to supress incomplete final line warning}
\usage{
read.table.nowarn(...)
}
\arguments{
\item{...}{arguments}
}
\description{
Function to supress incomplete final line warning
}
\details{
...
}

