% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{reduce}
\alias{reduce}
\title{reduce helper function to reduce data}
\usage{
reduce(data, year = NULL, fleet = NULL, age = NULL, conf = NULL)
}
\arguments{
\item{data}{a data object as returned by the function setup.sam.data}

\item{year}{a vector of years to be excluded.}

\item{fleet}{a vector of fleets to be excluded.}

\item{age}{a vector of ages fleets to be excluded.}

\item{conf}{an optional corresponding configuration to be modified along with the data change. Modified is returned as attribute "conf"}
}
\description{
reduce helper function to reduce data
}
\details{
When more than one vector is supplied they need to be of same length, as only the pairs are excluded
}
