% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMSY.R
\name{referencepoints}
\alias{referencepoints}
\alias{referencepoints.sam}
\title{Estimate reference points}
\usage{
referencepoints(fit, nYears, Fsequence, aveYears, selYears, catchType, ...)

\method{referencepoints}{sam}(fit, nYears = 100, Fsequence = seq(1e-05,
  4, len = 200), aveYears = as.numeric(c(length(fit$data$years) - 2:1)),
  selYears = as.numeric(c(length(fit$data$years) - 1)),
  catchType = "catch", ...)
}
\arguments{
\item{fit}{an object to calculate reference points for}

\item{...}{not used}
}
\value{
a sam_referencepoints fit
}
\description{
Work in progress - do not use
}
\author{
Christoffer Moesgaard Albertsen
}
