% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{residplot}
\alias{residplot}
\title{Compute and plot residual diagnostic tests for a SAM model fit}
\usage{
residplot(
  fit,
  resid = NULL,
  bubbles = TRUE,
  palette = hcl.colors(4, palette = "Blues"),
  p.breaks = c(0, 0.001, 0.05, 0.1, 1),
  reverseFleetOrder = FALSE,
  plot = TRUE
)
}
\arguments{
\item{fit}{an object of class ‘sam’}

\item{resid}{residuals (object of class "samres") from 'fit'. Calculated from 'fit' if not provided here.}

\item{bubbles}{show bubble plots (default: TRUE)}

\item{palette}{vector of colors used for p-value groupings}

\item{p.breaks}{breaks for p-value groupings}

\item{reverseFleetOrder}{reverse order of fleets within year for OSA residuals (default: FALSE)}

\item{plot}{plot pvalues (default: TRUE)}
}
\value{
list of matrices with p-values corresponding to the plots
}
\description{
Compute and plot residual diagnostic tests for a SAM model fit
}
\details{
Student's t-test for bias, Chi-square test for variance 1, Ljung-Box test on lag 1 for auto-correlation in time and age directions, F for equal variances for obs < or > median, and Shapiro-Wilk test for normality.
}
