% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.sam}
\alias{residuals.sam}
\title{Extract residuals from sam object}
\usage{
\method{residuals}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object (result from sam.fit)}

\item{...}{extra arguments}
}
\description{
Extract residuals from sam object
}
\details{
...
}

