% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmaxplot}
\alias{rmaxplot}
\alias{rmaxplot.default}
\alias{rmaxplot.samforecast}
\alias{rmaxplot.hcr}
\title{SAM rmax plot}
\usage{
rmaxplot(fit, ...)

\method{rmaxplot}{default}(fit, ...)

\method{rmaxplot}{samforecast}(fit, ...)

\method{rmaxplot}{hcr}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM rmax plot
}
\details{
Plot of life expectancy
}
