% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rmaxtable}
\alias{rmaxtable}
\alias{rmaxtable.default}
\title{rmax table}
\usage{
rmaxtable(fit, ...)

\method{rmaxtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
rmax table
}
\details{
...
}
