% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{rmvnorm}
\alias{rmvnorm}
\title{rmvnorm helper function to draw multivariate normal samples}
\usage{
rmvnorm(n = 1, mu, Sigma, pivot = FALSE, L)
}
\arguments{
\item{n}{the number of samples.}

\item{mu}{the mean vector.}

\item{Sigma}{a positive-definite symmetric matrix specifying the covariance matrix.}

\item{pivot}{Do pivot in chol?}
}
\value{
If n = 1 a vector of the same length as mu, otherwise an n by length(mu) matrix with one sample in each row.
}
\description{
rmvnorm helper function to draw multivariate normal samples
}
\details{
Generates samples via the Cholesky decomposition, which is less platform dependent than eigenvalue decomposition.
}
