% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{runwithout}
\alias{runwithout}
\title{runwithout helper function}
\usage{
runwithout(fit, year = NULL, fleet = NULL, map = fit$obj$env$map, ...)
}
\arguments{
\item{fit}{a fitted model object as returned from sam.fit}

\item{year}{a vector of years to be excluded.  When both fleet and year are supplied they need to be of same length, as only the pairs are excluded}

\item{fleet}{a vector of fleets to be excluded.  When both fleet and year are supplied they need to be of same length, as only the pairs are excluded}

\item{map}{map from original fit}

\item{...}{extra arguments to sam.fit}
}
\description{
runwithout helper function
}
\details{
...
}
