% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{sam.fit}
\alias{sam.fit}
\title{Fit SAM model}
\usage{
sam.fit(data, conf, parameters, newtonsteps = 0, rm.unidentified = FALSE,
  run = TRUE, lower = getLowerBounds(parameters),
  upper = getUpperBounds(parameters), sim.condRE = TRUE, ...)
}
\arguments{
\item{data}{data for the sam model as returned from the setup.sam.data function}

\item{conf}{model configuration which can be set up using the \code{\link{defcon}} function and then modified.}

\item{parameters}{initial values which can be set up using the \code{\link{defpar}} function and then modified.}

\item{newtonsteps}{optional extra true newton steps}

\item{rm.unidentified}{option to eliminate unidentified model parameters based on gradient in initial value (somewhat experimental)}

\item{run}{if FALSE return AD object without running the optimization}

\item{lower}{named list with lower bounds for optimization (only met before extra newton steps)}

\item{upper}{named list with upper bounds for optimization (only met before extra newton steps)}

\item{sim.condRE}{logical with default \code{TRUE}. Simulated observations will be conditional on estimated values of F and N, rather than also simulating F and N forward from their initial values.}

\item{...}{extra arguments to MakeADFun}
}
\value{
an object of class \code{sam}
}
\description{
Fit SAM model
}
\details{
...
}
\examples{
data(nscodData)
data(nscodConf)
data(nscodParameters)
fit <- sam.fit(nscodData, nscodConf, nscodParameters)
}

