% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{saveConf}
\alias{saveConf}
\title{Saves a model configuration list to a file}
\usage{
saveConf(x, file = "", overwrite = FALSE)
}
\arguments{
\item{x}{sam configuration list as returned from defcon or loadConf}

\item{file}{the file to save the configuration to}

\item{overwrite}{logical if an existing file should be overwritten (FALSE by default)}
}
\description{
Saves a model configuration list to a file
}
\details{
function useful for saving a model configuration. A saved configuration can be read back in via the loadConf function
}
