% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sprplot}
\alias{sprplot}
\alias{sprplot.default}
\alias{sprplot.samforecast}
\alias{sprplot.hcr}
\title{SAM SPR plot}
\usage{
sprplot(fit, ...)

\method{sprplot}{default}(fit, ...)

\method{sprplot}{samforecast}(fit, ...)

\method{sprplot}{hcr}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM SPR plot
}
\details{
Plot of deterministic equilibrium spawners per recruit assuming biological parameters and selectivity for that year remains unchanged in the future.
}
