% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sprtable}
\alias{sprtable}
\alias{sprtable.default}
\title{SPR table}
\usage{
sprtable(fit, ...)

\method{sprtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
SPR table
}
\details{
...
}
