% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{ssbplot}
\alias{ssbplot}
\title{SAM SSB plot}
\usage{
ssbplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM SSB plot
}
\details{
...
}

