% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_referencepoints.R
\name{stochasticReferencepoints}
\alias{stochasticReferencepoints}
\alias{stochasticReferencepoints.sam}
\title{Estimate stochastic reference points}
\usage{
stochasticReferencepoints(fit, referencepoints, ...)

\method{stochasticReferencepoints}{sam}(
  fit,
  referencepoints,
  method = "Median",
  catchType = "catch",
  nYears = 100,
  Frange = c(0, 2),
  nosim = 200,
  aveYears = max(fit$data$years) + (-4:0),
  selYears = max(fit$data$years) + (-4:0),
  newton.control = list(),
  seed = .timeToSeed(),
  knots = NA,
  nosim_ci = 200,
  derivedSummarizer = NA,
  nTail = 1,
  constraint = "F=\%f",
  deterministicF = FALSE,
  processNoiseF = FALSE,
  Fsequence = seq(min(Frange), max(Frange), len = 50),
  run = TRUE,
  DT = 0,
  equilibriumMethod = c("ES", "EC", "AD"),
  ncores = 1,
  tailSummarizer = mean,
  gridSummarizer = median,
  mc.type = "mclapply",
  ...
)
}
\arguments{
\item{fit}{a sam fit}

\item{referencepoints}{a character vector of reference points to estimate (see Details)}

\item{...}{additional parameters that can be passed on}

\item{method}{estimation method (See Details)}

\item{catchType}{catch type: catch, landing, discard}

\item{nYears}{Number of years to forecast}

\item{Frange}{Range of F values to consider}

\item{nosim}{Number of simulations for estimation}

\item{aveYears}{Years to average over for biological input}

\item{selYears}{Years to average over for selectivity}

\item{newton.control}{List of control parameters for optimization}

\item{seed}{Seed for simulations}

\item{knots}{Number of knots to use}

\item{nosim_ci}{Number of simulations for bootstrap confidence intervals}

\item{derivedSummarizer}{Function to summarize derived per-recruit values}

\item{nTail}{Number of years from the simulation to include in calculations}

\item{constraint}{Format of forecast constraint. "\%f" is replaced by F values.}

\item{deterministicF}{If FALSE, modelled logF process noise will be added to target logF in forecasts.}

\item{Fsequence}{F sequence to explore}

\item{run}{run it?}

\item{DT}{...}

\item{equilibriumMethod}{method to use}

\item{ncores}{Number of cores}
}
\value{
reference point object
}
\description{
The function estimates reference points based on stochastic model forecasts.
}
\details{
The following reference points are implemented:
\describe{
   \item{F=x}{F fixed to x, e.g., \code{"F=0.3"} (NOT IMPLEMENTED YET)}
   \item{StatusQuo}{F in the last year of the assessment (NOT IMPLEMENTED YET)}
   \item{StatusQuo-y}{F in the y years before the last in the assessment, e.g., \code{"StatusQuo-1"} (NOT IMPLEMENTED YET)}
   \item{MSY}{F that maximizes yield}
   \item{0.xMSY}{Fs that gives 0.x*100\% of MSY, e.g., \code{"0.95MSY"}}
   \item{Max}{F that maximizes yield per recruit (NOT IMPLEMENTED YET)}
   \item{0.xdYPR}{F such that the derivative of yield per recruit is 0.x times the derivative at F=0, e.g., \code{"0.1dYPR"} (NOT IMPLEMENTED YET)}
   \item{0.xSPR}{F such that spawners per recruit is 0.x times spawners per recruit at F=0, e.g., \code{"0.35SPR"} (NOT IMPLEMENTED YET)}
   \item{0.xB0}{F such that biomass is 0.x times the biomass at F=0, e.g., \code{"0.2B0"} (NOT IMPLEMENTED YET)}
}
Reference points can be estimated using these methods:
\describe{
   \item{Mean}{Use least squares to estimate mean equilibrium values}
   \item{Q0.x}{Use quantile regression to estimate the 0.x quantile of equilibrium values}
   \item{Median}{Identical to Q0.5}
   \item{Mode}{(NOT IMPLEMENTED YET)}
}
To estimate median equilibrium yield, as required by ICES, the method "Q0.5" should be used.
Note that this function is highly experimental.
}
\examples{
\dontrun{
  stochasticReferencepoints(fit, c("MSY","0.95MSY"))
}
}
