% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{stockassessment-deprecated}
\alias{stockassessment-deprecated}
\title{Deprecated and defunct functions}
\description{
Deprecated and defunct functions
}
\section{\code{referencepoints}}{

For \code{referencepoints}, use \code{\link{deterministicReferencepoints}}.
}

