% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.sam}
\alias{summary.sam}
\title{Summary of sam object}
\usage{
\method{summary}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object (result from sam.fit)}

\item{...}{extra arguments}
}
\description{
Summary of sam object
}
\details{
...
}

