% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{tsbplot}
\alias{tsbplot}
\title{SAM TSB plot}
\usage{
tsbplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM TSB plot
}
\details{
...
}

