% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.data.files}
\alias{write.data.files}
\title{Write all data files from a model object in the usual ICES/CEFAS data files}
\usage{
write.data.files(fit, dir = ".")
}
\arguments{
\item{fit}{A fitted object as returned from sam.fit}

\item{dir}{Directory where the files are written}
}
\description{
Write all data files from a model object in the usual ICES/CEFAS data files
}
\details{
Write all data files from a model object in the usual ICES/CEFAS data files
}
