% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.ices}
\alias{write.ices}
\title{Write ICES/CEFAS data file from matrix}
\usage{
write.ices(x, fileout, writeToOne = TRUE, ...)
}
\arguments{
\item{x}{a matrix where rownames are taken as years and colnames are taken as ages}

\item{fileout}{file name or connection}

\item{writeToOne}{Write multi fleet data to one file if data is equal for all fleets}

\item{...}{Arguments to be passed to write}
}
\description{
Write ICES/CEFAS data file from matrix
}
\details{
Takes the data and writes them in the ICES/CEFAS format. It is assumed that rows represent consecutive years and cols consecutive ages
}
