% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.surveys}
\alias{write.surveys}
\title{Write surveys in ICES/CEFAS data file from a model object}
\usage{
write.surveys(fit, fileout, ...)
}
\arguments{
\item{fit}{A fitted object as returned from sam.fit}

\item{fileout}{file name or connection}

\item{...}{Arguments to be passed to write}
}
\description{
Write surveys in ICES/CEFAS data file from a model object
}
\details{
Takes the survey data from the fitted object and writes them in the ICES/CEFAS format.
}
