% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{yearslosttable}
\alias{yearslosttable}
\alias{yearslosttable.default}
\title{Years Lost table}
\usage{
yearslosttable(fit, cause, ...)

\method{yearslosttable}{default}(fit, cause = c("Fishing", "Other", "LifeExpectancy"), ...)
}
\arguments{
\item{fit}{...}

\item{cause}{Fisning, Other, or LifeExpectancy}

\item{...}{extra arguments not currently used}
}
\description{
Years Lost table
}
\details{
...
}
