% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yprplot}
\alias{yprplot}
\title{Plots the classic yield per recruit}
\usage{
yprplot(fit, Flimit = 2, Fdelta = 0.01, aveYears = 15, ageLimit = 100)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{Flimit}{Upper limit for Fbar}

\item{Fdelta}{increments on the Fbar axis}

\item{aveYears}{Number of years back to use when calculating averages (selection, weights, ...)}

\item{ageLimit}{Oldest age used (should be high)}
}
\description{
Plots the classic yield per recruit
}
