% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{yprtable}
\alias{yprtable}
\alias{yprtable.default}
\title{YPR table}
\usage{
yprtable(fit, ...)

\method{yprtable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
YPR table
}
\details{
...
}
