% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{catchplot}
\alias{catchplot}
\title{SAM catch plot}
\usage{
catchplot(fit, obs.show = TRUE, drop = NULL, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{obs.show}{if observations are to be shown also}

\item{drop}{number of years to be left unplotted at the end. Default (NULL) is to not show years at the end with no catch information}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM catch plot
}
\details{
...
}

