% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{catchplot}
\alias{catchplot}
\alias{catchplot.sam}
\alias{catchplot.samset}
\alias{catchplot.samforecast}
\title{SAM catch plot}
\usage{
catchplot(fit, obs.show = TRUE, drop = NULL, ...)

\method{catchplot}{sam}(fit, obs.show = TRUE, drop = NULL, plot = TRUE, ...)

\method{catchplot}{samset}(fit, obs.show = TRUE, drop = NULL, ...)

\method{catchplot}{samforecast}(fit, obs.show = TRUE, drop = NULL, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{obs.show}{if observations are to be shown also}

\item{drop}{number of years to be left unplotted at the end. Default (NULL) is to not show years at the end with no catch information}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}

\item{plot}{true if catch should be plotted}
}
\description{
SAM catch plot
}
\details{
Plot of estimated (and optionally observed) total catch in weight
}
