% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catchtable}
\alias{catchtable}
\title{Catch table}
\usage{
catchtable(fit, obs.show = FALSE)
}
\arguments{
\item{fit}{object returned from sam.fit}

\item{obs.show}{logical add a column with catch sum of product rowsums(C*W)}
}
\description{
Catch table
}
\details{
...
}
