% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{defcon}
\alias{defcon}
\title{Setup basic minimal configuration for sam assessment}
\usage{
defcon(dat)
}
\arguments{
\item{dat}{sam data object}
}
\description{
Setup basic minimal configuration for sam assessment
}
\details{
...
}

