% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\name{defpar}
\alias{defpar}
\title{Setup minimal initial parameters}
\usage{
defpar(dat, conf)
}
\arguments{
\item{dat}{sam data object}

\item{conf}{configuration list}
}
\description{
Setup minimal initial parameters
}
\details{
...
}

