% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\name{defpar}
\alias{defpar}
\title{Setup minimal initial parameters}
\usage{
defpar(dat, conf)
}
\arguments{
\item{dat}{sam data object}

\item{conf}{configuration list}
}
\value{
a list containing the following
\item{logFpar}{}
\item{logQpow}{}
\item{logSdLogFsta}{}
\item{logSdLogN}{}
\item{logSdLogObs}{}
\item{logSdLogTotalObs}{}
\item{transfIRARdist}{}
\item{sigmaObsParUS}{}
\item{rec_loga}{}
\item{rec_logb}{}
\item{itrans_rho}{}
\item{logScale}{}
\item{logScaleSSB}{}
\item{logPowSSB}{}
\item{logSdSSB}{}
\item{logitReleaseSurvival}{}
\item{logitRecapturePhi}{}
\item{logF}{}
\item{logN}{}
}
\description{
Setup minimal initial parameters
}
\details{
...
}

