% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_referencepoints.R
\name{deterministicReferencepoints}
\alias{deterministicReferencepoints}
\title{Function to calculate reference points for the embedded deterministic model of a SAM fit}
\usage{
deterministicReferencepoints(fit, referencepoints, ...)
}
\arguments{
\item{fit}{A fitted SAM model}

\item{referencepoints}{list of reference points to calculate (See details)}

\item{...}{other arguments not used}
}
\value{

}
\description{
Function to calculate reference points for the embedded deterministic model of a SAM fit
}
\author{
Christoffer Moesgaard Albertsen
}
