% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{fbarplot}
\alias{fbarplot}
\title{SAM Fbar plot}
\usage{
fbarplot(fit, partial = TRUE, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{partial}{true if included partial F's are to be plotted}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM Fbar plot
}
\details{
...
}

