% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{fbarplot}
\alias{fbarplot}
\title{SAM Fbar plot}
\usage{
fbarplot(fit, partial = (class(fit) == "sam"), drop = NULL, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{partial}{true if included partial F's are to be plotted}

\item{drop}{number of years to be left unplotted at the end. Default (NULL) is to not show years at the end with no catch information}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM Fbar plot
}
\details{
...
}

