% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fbarplot}
\alias{fbarplot}
\alias{fbarplot.sam}
\alias{fbarplot.samset}
\alias{fbarplot.samforecast}
\alias{fbarplot.hcr}
\title{SAM Fbar plot}
\usage{
fbarplot(fit, ...)

\method{fbarplot}{sam}(
  fit,
  partial = TRUE,
  drop = NULL,
  pcol = "lightblue",
  page = NULL,
  plot = TRUE,
  effectiveF = any(!fit$conf$seasonTimes \%in\% c(0, 1)),
  ...
)

\method{fbarplot}{samset}(
  fit,
  partial = FALSE,
  drop = NULL,
  pcol = "lightblue",
  page = NULL,
  ...
)

\method{fbarplot}{samforecast}(
  fit,
  partial = FALSE,
  drop = NULL,
  pcol = "lightblue",
  page = NULL,
  ...
)

\method{fbarplot}{hcr}(
  fit,
  partial = FALSE,
  drop = NULL,
  pcol = "lightblue",
  page = NULL,
  ...
)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}

\item{partial}{true if included partial F's are to be plotted}

\item{drop}{number of years to be left unplotted at the end. Default (NULL) is to not show years at the end with no catch information}

\item{pcol}{color of partial lines}

\item{page}{partial ages to plot}

\item{plot}{true if fbar should be plotted}

\item{effectiveF}{If TRUE, effective full year F based on catch and survival is plotted. If FALSE, full year F based on survival is plotted.}
}
\description{
SAM Fbar plot
}
\details{
Plot the defined fbar.
}
