% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{fitfromweb}
\alias{fitfromweb}
\title{Read a fitted model from stockassessment.org}
\usage{
fitfromweb(stockname)
}
\arguments{
\item{stockname}{The sort-form name of a stock on stockassessment.org. This will (currently?) not for stock defined via the AD Model builder version of SAM.}
}
\description{
Read a fitted model from stockassessment.org
}
\details{
...
}
