% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fitplot}
\alias{fitplot}
\title{Plots fit to data}
\usage{
fitplot(fit, log = TRUE, fleets = unique(fit$data$aux[, "fleet"]), ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{log}{should the plot be against log-obs}

\item{fleets}{an integer vector of fleets to plot. Default is all of them}

\item{...}{extra arguments to plot}
}
\description{
Plots fit to data
}
