% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{forecastMSY}
\alias{forecastMSY}
\alias{forecastMSY.sam}
\title{Estimating Fmsy}
\usage{
forecastMSY(fit, nYears = 100, nlminb.control = list(eval.max = 2000,
  iter.max = 2000), rec.years = c(), processNoiseF = FALSE, ...)

\method{forecastMSY}{sam}(fit, nYears = 100,
  nlminb.control = list(eval.max = 2000, iter.max = 2000, trace = 1),
  rec.years = c(), processNoiseF = FALSE, ...)
}
\arguments{
\item{fit}{a SAM fit}

\item{nYears}{Number of years to forecast}

\item{nlminb.control}{list of control variables for nlminb}

\item{rec.years}{Numeric vector of years to use (to calculate mean and standard deviation) for recruitment. An empty vector will use the recruitment model.}

\item{processNoiseF}{Should random walk process noise be used for F?}

\item{...}{other arguments}
}
\description{
Work in progress - do not use
}
\seealso{
forecast
}
\author{
Christoffer Moesgaard Albertsen
}
