% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{forecastMSY}
\alias{forecastMSY}
\alias{forecastMSY.sam}
\title{Estimating Fmsy}
\usage{
forecastMSY(fit, nYears = 100, nlminb.control = list(eval.max = 2000,
  iter.max = 2000), rec.years = c(), ave.years = max(fit$data$years) +
  (-9:0), processNoiseF = FALSE, ...)

\method{forecastMSY}{sam}(fit, nYears = 100,
  nlminb.control = list(eval.max = 2000, iter.max = 2000, trace = 1),
  rec.years = c(), ave.years = max(fit$data$years) + (-9:0),
  processNoiseF = FALSE, jacobianHScale = 0.5,
  nCatchAverageYears = 20, ...)
}
\arguments{
\item{fit}{a SAM fit}

\item{nYears}{Number of years to forecast}

\item{nlminb.control}{list of control variables for nlminb}

\item{rec.years}{Numeric vector of years to use (to calculate mean and standard deviation) for recruitment. An empty vector will use the recruitment model.}

\item{ave.years}{vector of years to average for weights, maturity, M and such. Following ICES guidelines, the default is the last 10 years.}

\item{processNoiseF}{Should random walk process noise be used for F?}

\item{...}{other arguments passed to forecast}

\item{jacobianHScale}{Scale step size in jacobian calculation}

\item{nCatchAverageYears}{Number of years to average catch over for finding MSY}
}
\description{
Estimating Fmsy
}
\references{
Albertsen, C. M. and Trijoulet, V. (2020) Model-based estimates of reference points in an age-based state-space stock assessment model. Fisheries Research, 230, 105618. doi: 10.1016/j.fishres.2020.105618
}
\seealso{
\link{forecast} \link{referencepoints}
}
\author{
Christoffer Moesgaard Albertsen
}
