% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fselectivityplot}
\alias{fselectivityplot}
\alias{fselectivityplot.sam}
\title{SAM F-selectivity plot}
\usage{
fselectivityplot(fit, cexAge = 1, ...)

\method{fselectivityplot}{sam}(fit, cexAge = 1, ...)
}
\arguments{
\item{fit}{An object returned from sam.fit}

\item{cexAge}{cex variable giving the size of the age numbers}

\item{...}{extra arguments transferred to barplot and text}
}
\description{
SAM F-selectivity plot
}
\details{
Plots selectivity in F.
}
