% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fSelectivityPlot}
\alias{fSelectivityPlot}
\alias{fSelectivityPlot.sam}
\title{SAM F-selectivity plot}
\usage{
fSelectivityPlot(fit, ...)

\method{fSelectivityPlot}{sam}(fit, ...)
}
\arguments{
\item{fit}{An object returned from sam.fit}

\item{...}{extra arguments transferred to barplot}
}
\description{
SAM F-selectivity plot
}
\details{
Plots selectivity in F.
}
