% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{grad}
\alias{grad}
\title{Calculate gradient of a function}
\usage{
grad(func, x, ...)
}
\arguments{
\item{func}{function}

\item{x}{parameter values}

\item{...}{passed to func}
}
\value{
gradient vector
}
\description{
Calculate gradient of a function
}
\author{
Christoffer Moesgaard Albertsen
}
