% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr}
\alias{hcr}
\alias{hcr.sam}
\title{Harvest control rule forecast}
\usage{
hcr(fit, ...)

\method{hcr}{sam}(
  fit,
  nYears = 20,
  Ftarget,
  Btrigger,
  Flim = 0,
  Blim = 0,
  Flow = 0,
  Blow = 0,
  nosim = 10000,
  ave.years = max(fit$data$years) + (-4:0),
  rec.years = c(),
  preForecast = list(),
  ...
)
}
\arguments{
\item{fit}{A SAM fit}

\item{nYears}{Number of years to forecast}

\item{Ftarget}{Target F for high SSB}

\item{Btrigger}{SSB that triggers the control rule}

\item{Flim}{F used for SSB = Blim}

\item{Blim}{Between Blim and Btrigger, F values are selected based on linear interpolation from Flim to Ftarget}

\item{Flow}{F used for SSB < Blow}

\item{Blow}{SSB for which Flow is used below}

\item{nosim}{Number of simulations to do. If NULL a model forecast based on the Laplace approximation is used}

\item{ave.years}{vector of years to average for weights, maturity, M and such}

\item{rec.years}{vector of years to use to resample recruitment from. If an empty vector is given, recruitment is based on the fitted model.}

\item{preForecast}{list of forecast parameters (i.e., fval, fscale, catchval, landval, or nextssb) to use before the HCR}

\item{...}{additional arguments passed to \link{modelforecast}}
}
\value{
hcr object
}
\description{
The formula below is used to determine a new F based on the previous SSB.
\deqn{F = \left\{
              \begin{array}{ll}
                  F_{low} & SSB < B_{low} \\
                  min\left(Ftarget, \max\left( F_{lim}, (SSB - B_{lim}) \cdot (F_{target} - F_{lim}) / (B_{trigger}-B_{lim}) \right)\right) & SSB \ge B_{low}
              \end{array}\right.
}
}
\seealso{
modelforecast
}
\author{
Christoffer Moesgaard Albertsen
}
