% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{icesAdviceRule}
\alias{icesAdviceRule}
\title{Forecast with an ICES advice rule}
\usage{
icesAdviceRule(
  x,
  Fmsy,
  MSYBtrigger,
  Blim,
  nYears = 1,
  nosim = 10000,
  ave.years = max(fit$data$years) + (-4:0),
  rec.years = numeric(0),
  preForecast = list(),
  currentSSB = FALSE,
  ...
)
}
\arguments{
\item{x}{Fitted assessment model}

\item{Fmsy}{}

\item{MSYBtrigger}{}

\item{Blim}{}

\item{nosim}{}

\item{ave.years}{}

\item{rec.years}{}

\item{...}{}
}
\value{

}
\description{
Forecast with an ICES advice rule
}
\references{
ICES (2021) Advice on fishing opportunities. DOI: 10.17895/ices.advice.7720
}
\author{
Christoffer Moesgaard Albertsen
}
