% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{jacobian}
\alias{jacobian}
\title{Calculate jacobian of a function}
\usage{
jacobian(func, x, ...)
}
\arguments{
\item{func}{function}

\item{x}{parameter values}

\item{...}{passed to func}
}
\value{
jacobian matrix
}
\description{
Calculate jacobian of a function
}
\author{
Christoffer Moesgaard Albertsen
}
