% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{jacobian}
\alias{jacobian}
\title{Calculate jacobian of a function}
\usage{
jacobian(
  func,
  x,
  h = abs(1e-04 * x) + 1e-04 * (abs(x) < sqrt(.Machine$double.eps/7e-07)),
  ...
)
}
\arguments{
\item{func}{function}

\item{x}{parameter values}

\item{h}{step size}

\item{...}{passed to func}
}
\value{
jacobian matrix
}
\description{
Calculate jacobian of a function
}
