% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{leaveout}
\alias{leaveout}
\title{leaveout run}
\usage{
leaveout(fit, fleet = as.list(2:fit$data$noFleets), ncores = detectCores(),
  ...)
}
\arguments{
\item{fit}{a fitted model object as returned from sam.fit}

\item{fleet}{a list of vectors. Each element in the list specifies a run where the fleets mentioned are omitted}

\item{ncores}{the number of cores to attemp to use}

\item{...}{extra arguments to \code{\link{sam.fit}}}
}
\description{
leaveout run
}
\details{
...
}
