% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lifeexpectancyplot}
\alias{lifeexpectancyplot}
\alias{lifeexpectancyplot.default}
\alias{lifeexpectancyplot.samforecast}
\alias{lifeexpectancyplot.hcr}
\title{SAM life expectancy plot}
\usage{
lifeexpectancyplot(fit, atRecruit = TRUE, ...)

\method{lifeexpectancyplot}{default}(fit, atRecruit = TRUE, ylimAdd = fit$conf$maxAge, ...)

\method{lifeexpectancyplot}{samforecast}(fit, atRecruit = TRUE, ylimAdd = fit$conf$maxAge, ...)

\method{lifeexpectancyplot}{hcr}(fit, atRecruit = TRUE, ylimAdd = fit$conf$maxAge, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{atRecruit}{If true, show life expectancy given survival until minAge, otherwise show life expectancy at birth}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}

\item{ylimAdd}{values to add when calculating ylim for the plot}
}
\description{
SAM life expectancy plot
}
\details{
Plot of life expectancy
}
