% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{lifeexpectancytable}
\alias{lifeexpectancytable}
\alias{lifeexpectancytable.default}
\title{Life expectancy table}
\usage{
lifeexpectancytable(fit, atRecruit = TRUE, ...)

\method{lifeexpectancytable}{default}(fit, atRecruit = TRUE, ...)
}
\arguments{
\item{fit}{...}

\item{atRecruit}{If true, show life expectancy given survival until minAge, otherwise show life expectancy at birth}

\item{...}{extra arguments not currently used}
}
\description{
Life expectancy table
}
\details{
...
}
