% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.sam}
\alias{logLik.sam}
\title{Log likelihood of sam object}
\usage{
\method{logLik}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object (result from sam.fit)}

\item{...}{extra arguments}
}
\description{
Log likelihood of sam object
}
\details{
...
}

