% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{modeltable}
\alias{modeltable}
\title{model table}
\usage{
modeltable(fits)
}
\arguments{
\item{fits}{A sam fit as returned from the sam.fit function, or a collection c(fit1, fit2, ...) of such fits}
}
\description{
model table
}
\details{
...
}
