% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{obscorrplot}
\alias{obscorrplot}
\title{Plots the estimated correlation matrices by fleet.}
\usage{
obscorrplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments to plot}
}
\description{
Plots the estimated correlation matrices by fleet.
}

