% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{obscov}
\alias{obscov}
\title{Extract observation covariance matrices from a SAM fit}
\usage{
obscov(fit, corr = FALSE)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{corr}{if TRUE return correlation matrices rather than covariances}
}
\value{
a list of matrices
}
\description{
Extract observation covariance matrices from a SAM fit
}
