% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{parplot}
\alias{parplot}
\alias{parplot.sam}
\alias{parplot.samset}
\title{SAM parameter plot}
\usage{
parplot(fit, cor.report.limit = 0.95, ...)

\method{parplot}{sam}(fit, cor.report.limit = 0.95, ...)

\method{parplot}{samset}(fit, cor.report.limit = 0.95, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{cor.report.limit}{correlations with absolute value > this number is reported in the plot}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM parameter plot
}
\details{
Plot of all estimated model parameters (fixed effects). Shown with confidence interval.
}
