% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{parplot}
\alias{parplot}
\title{SAM parameter plot}
\usage{
parplot(fit, cor.report.limit = 0.95, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{cor.report.limit}{correlations with absolute value > this number is reported in the plot}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM parameter plot
}
\details{
...
}

