% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.sam}
\alias{plot.sam}
\title{Plot sam object}
\usage{
\method{plot}{sam}(x, ...)
}
\arguments{
\item{x}{fitted object as returned from the \code{\link{sam.fit}} function.}

\item{...}{extra arguments (not possible to use add=TRUE --- instead collect to a list of fits using e.g the c(...), and then plot that collected object).}
}
\description{
Plot sam object
}
\details{
gives a 3 plot overview plot og ssb, fbar, and recruits. These plots are available individually via the functions \code{\link{ssbplot}}, \code{fbarplot}, and \code{\link{recplot}}.
}
