% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.samres}
\alias{plot.samres}
\title{Plot sam residuals}
\usage{
\method{plot}{samres}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot sam residuals
}
\details{
...
}
\examples{
\dontrun{
data(nscodData)
data(nscodConf)
data(nscodParameters)
fit <- sam.fit(nscodData, nscodConf, nscodParameters)
plot(residuals(fit))
}
}

