% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.samres}
\alias{plot.samres}
\title{Plot sam residuals}
\usage{
\method{plot}{samres}(x, type = "bubble", ...)
}
\arguments{
\item{x}{an object of type 'samres' as returned from functions \code{\link{residuals.sam}} or \code{\link{procres}}.}

\item{type}{either "bubble" (default) or "summary"}

\item{...}{extra arguments}
}
\description{
Plot sam residuals
}
\details{
In the "bubble" type red indicate negative residuals and blue positive. The area of the circles scales with the absolute size of the residuals.
}
\examples{
\dontrun{
data(nscodData)
data(nscodConf)
data(nscodParameters)
fit <- sam.fit(nscodData, nscodConf, nscodParameters)
par(ask=FALSE)
plot(residuals(fit))
}
}
