% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.samset}
\alias{plot.samset}
\title{Plot sam object}
\usage{
\method{plot}{samset}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot sam object
}
\details{
...
}

