% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotby}
\alias{plotby}
\title{Plot by one or two}
\usage{
plotby(x = NULL, y = NULL, z = NULL, by = NULL, bubblescale = 1,
  x.common = TRUE, y.common = TRUE, xlab = NULL, ylab = NULL,
  xlim = NULL, ylim = NULL, axes = TRUE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{z}{numeric vector}

\item{by}{vector or two column matrix to create sub sets from}

\item{bubblescale}{scaling of bubble size}

\item{x.common}{logical: use same x-axis for all plots}

\item{y.common}{logical: use same y-axis for all plots}

\item{xlab}{normal graphical parameter}

\item{ylab}{normal graphical parameter}

\item{xlim}{normal graphical parameter}

\item{ylim}{normal graphical parameter}

\item{axes}{normal graphical parameter}

\item{...}{additional graphical parameters}
}
\description{
Plot by one or two
}
\details{
...
}

