% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{predstdplot}
\alias{predstdplot}
\alias{predstdplot.sam}
\title{Plot prediction-standard deviation relation}
\usage{
predstdplot(fit, fleet, age = NULL, type = "log", ylim = NULL, ...)

\method{predstdplot}{sam}(fit, fleet, age = NULL, type = "log", ylim = NULL, ...)
}
\arguments{
\item{fit}{An object returned from sam.fit}

\item{fleet}{Fleet to plot relation for}

\item{age}{Age to plot relation for. Variable not used if there is only one relation in the feet.}

\item{type}{Either "log" or "natural". Relation shown on log or natural scale.}

\item{ylim}{range of y-axis, default is set to match confidence intervals.}

\item{...}{extra arguments transferred to plot and lines}
}
\description{
Plot prediction-standard deviation relation
}
\details{
Plots relation between prediction and standard deviation.
}
