% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sam}
\alias{print.sam}
\title{Print sam object}
\usage{
\method{print}{sam}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print sam object
}
\details{
...
}

