% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samcoef}
\alias{print.samcoef}
\title{Print samcoef object}
\usage{
\method{print}{samcoef}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print samcoef object
}
\details{
...
}

