% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{procres}
\alias{procres}
\title{Compute process residuals (single joint sample)}
\usage{
procres(fit, map = fit$obj$env$map, ...)
}
\arguments{
\item{fit}{the fitted object as returned from the \code{\link{sam.fit}} function}

\item{map}{map from original fit}

\item{...}{extra arguments (not currently used)}
}
\value{
an object of class \code{samres}
}
\description{
Compute process residuals (single joint sample)
}
\details{
Single joint sample residuals of log(F) and log(N)
}
