% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{recplot}
\alias{recplot}
\title{SAM Recruits plot}
\usage{
recplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot}
}
\description{
SAM Recruits plot
}
\details{
...
}

