% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{recplot}
\alias{recplot}
\title{SAM Recruits plot}
\usage{
recplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM Recruits plot
}
\details{
Plot of numbers of recruits (youngest age class)
}
