% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMSY.R
\name{referencepoints}
\alias{referencepoints}
\alias{referencepoints.sam}
\title{Estimate reference points}
\usage{
referencepoints(fit, ...)

\method{referencepoints}{sam}(fit, ...)
}
\arguments{
\item{fit}{an object to calculate reference points for}

\item{...}{not used}
}
\value{
a sam_referencepoints fit
}
\description{
Work in progress - do not use
}
\author{
Christoffer Moesgaard Albertsen
}
