% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{referencepoints}
\alias{referencepoints}
\alias{referencepoints.sam}
\title{Estimate reference points}
\usage{
referencepoints(fit, nYears, Fsequence, aveYears, selYears, catchType, ...)

\method{referencepoints}{sam}(fit, nYears = 100, Fsequence = seq(1e-05,
  4, len = 200), aveYears = max(fit$data$years) + (-4:0),
  selYears = max(fit$data$years), SPRpercent = c(0.35),
  catchType = "catch", ...)
}
\arguments{
\item{fit}{an object to calculate reference points for}

\item{nYears}{Number of years to use in per-recruit calculations}

\item{Fsequence}{Sequence of F values used to report per-recruit and equilibrium values}

\item{aveYears}{Vector of year indices used to calculate average natural mortality, weights, etc. (starting at 0)}

\item{selYears}{Vector of year indices used to calculate selectivity (starting at 0)}

\item{catchType}{Catch type used: (total) catch, landings, discard.}

\item{...}{not used}
}
\value{
a sam_referencepoints fit
}
\description{
Work in progress - do not use
}
\author{
Christoffer Moesgaard Albertsen
}
