% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R
\name{referencepoints}
\alias{referencepoints}
\alias{referencepoints.sam}
\title{Estimate reference points}
\usage{
referencepoints(fit, nYears, Fsequence, aveYears, selYears, SPRpercent,
  catchType, MSYreduction, newtonSteps = 3, optN = 100, ...)

\method{referencepoints}{sam}(fit, nYears = 100, Fsequence = seq(0, 4,
  len = 200), aveYears = max(fit$data$years) + (-9:0),
  selYears = max(fit$data$years), SPRpercent = c(0.35),
  catchType = "catch", jacobianHScale = 0.5, MSYreduction = c(0.05),
  newtonSteps = 3, optN = 100, ...)
}
\arguments{
\item{fit}{an object to calculate reference points for}

\item{nYears}{Number of years to use in per-recruit calculations}

\item{Fsequence}{Sequence of F values used to report per-recruit and equilibrium values}

\item{aveYears}{Vector of year indices used to calculate average natural mortality, weights, etc. Following ICES guidelines, the default is the last 10 years (starting at 0)}

\item{selYears}{Vector of year indices used to calculate selectivity (starting at 0)}

\item{SPRpercent}{Vector of x values for F[x * 100\%] reference points. Default is 0.35.}

\item{catchType}{Catch type used: (total) catch, landings, discard.}

\item{MSYreduction}{Vector of proportions for MSY ranges. Default is 0.05 giving an MSY range corresponding to no more than a 5% yield reduction.}

\item{newtonSteps}{Number of additional Newton steps at the end of the reference point optimization.}

\item{...}{not used}
}
\value{
a sam_referencepoints fit
}
\description{
Work in progress - do not use
}
\references{
Albertsen, C. M. and Trijoulet, V. (2020) Model-based estimates of reference points in an age-based state-space stock assessment model. Fisheries Research, 230, 105618. doi: 10.1016/j.fishres.2020.105618
}
\seealso{
\link{forecastMSY}
}
\author{
Christoffer Moesgaard Albertsen
}
