% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.sam}
\alias{residuals.sam}
\title{Extract residuals from sam object}
\usage{
\method{residuals}{sam}(object, discrete = FALSE, ...)
}
\arguments{
\item{object}{sam fitted object (result from sam.fit)}

\item{discrete}{logical if model contain discrete observations}

\item{...}{extra arguments for TMB's oneStepPredict}
}
\description{
Extract residuals from sam object
}
\details{
...
}

