% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{sam.fit}
\alias{sam.fit}
\title{Fit SAM model}
\usage{
sam.fit(data, conf, parameters, ...)
}
\arguments{
\item{data}{...}

\item{conf}{...}

\item{parameters}{...}

\item{...}{extra arguments to MakeADFun}
}
\description{
Fit SAM model
}
\details{
...
}
\examples{
data(nscodData)
data(nscodConf)
data(nscodParameters)
fit <- sam.fit(nscodData, nscodConf, nscodParameters)
}

