% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{sam.fit}
\alias{sam.fit}
\title{Fit SAM model}
\usage{
sam.fit(data, conf, parameters, newtonsteps = 3, rm.unidentified = FALSE,
  run = TRUE, lower = getLowerBounds(parameters),
  upper = getUpperBounds(parameters), ...)
}
\arguments{
\item{data}{data for the sam model as returned from the setup.sam.data function}

\item{conf}{model configuration for the model in a format similar to what is returned from the defcon function}

\item{parameters}{initial values for the model in a format similar to what is returned from the defpar function}

\item{newtonsteps}{optional extra true extra newton steps}

\item{rm.unidentified}{option to eliminate unidentified model parameters based on gradient in initial value (somewhat experimental)}

\item{run}{if FALSE return AD object without running the optimization}

\item{lower}{named list with lower bounds for optimization (only met before extra newton steps)}

\item{upper}{named list with upper bounds for optimization (only met before extra newton steps)}

\item{...}{extra arguments to MakeADFun}
}
\description{
Fit SAM model
}
\details{
...
}
\examples{
data(nscodData)
data(nscodConf)
data(nscodParameters)
fit <- sam.fit(nscodData, nscodConf, nscodParameters)
}

