% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sdplot}
\alias{sdplot}
\alias{sdplot.sam}
\title{Plots the sd of the log observations as estimated in SAM in increasing order}
\usage{
sdplot(fit, barcol = NULL, marg = NULL, ylim = NULL, ...)

\method{sdplot}{sam}(fit, barcol = NULL, marg = NULL, ylim = NULL, show.rel.w = FALSE, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{barcol}{color for each fleet and age}

\item{marg}{margin for plot (mar in par())}

\item{ylim}{bounds for y-axis}

\item{...}{extra arguments to plot}

\item{show.rel.w}{plots the relative weight of each observation rather than the sd, estimated as (1/sd^2)/max(1/sd^2)}
}
\description{
Plots the sd of the log observations as estimated in SAM in increasing order
}
