% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{setup.sam.data}
\alias{setup.sam.data}
\title{Combine the data sources to SAM readable object}
\usage{
setup.sam.data(
  fleets = NULL,
  surveys = NULL,
  residual.fleet = NULL,
  prop.mature = NULL,
  stock.mean.weight = NULL,
  catch.mean.weight = NULL,
  dis.mean.weight = NULL,
  land.mean.weight = NULL,
  natural.mortality = NULL,
  prop.f = NULL,
  prop.m = NULL,
  land.frac = NULL,
  recapture = NULL,
  agesampledata = NULL
)
}
\arguments{
\item{fleets}{comm fleets vith effort (currently unimplemented)}

\item{surveys}{surveys}

\item{residual.fleet}{total catch minus commercial}

\item{prop.mature}{pm}

\item{stock.mean.weight}{sw}

\item{catch.mean.weight}{cw}

\item{dis.mean.weight}{dw}

\item{land.mean.weight}{lw}

\item{natural.mortality}{nm}

\item{prop.f}{...}

\item{prop.m}{...}

\item{land.frac}{...}

\item{recapture}{...}

\item{agesampledata}{A data frame with 3 columns age, sample, and data corresponding to an age reading study where an age was determined for a sample within a data set. If only one dataset is available, then the data coulmn can just be the same integer in all rows.}
}
\description{
Combine the data sources to SAM readable object
}
\details{
...
}
