% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{setup.sam.data}
\alias{setup.sam.data}
\title{Combine the data sources to SAM readable object}
\usage{
setup.sam.data(
  fleets = NULL,
  surveys = NULL,
  residual.fleets = NULL,
  prop.mature = NULL,
  stock.mean.weight = NULL,
  catch.mean.weight = NULL,
  dis.mean.weight = NULL,
  land.mean.weight = NULL,
  natural.mortality = NULL,
  prop.f = NULL,
  prop.m = NULL,
  land.frac = NULL,
  recapture = NULL,
  sum.residual.fleets = NULL,
  keep.all.ages = FALSE
)
}
\arguments{
\item{fleets}{comm fleets vith effort (currently unimplemented)}

\item{surveys}{surveys}

\item{residual.fleets}{fleet, or list of fleets without effort information}

\item{prop.mature}{pm}

\item{stock.mean.weight}{sw}

\item{catch.mean.weight}{cw}

\item{dis.mean.weight}{dw}

\item{land.mean.weight}{lw}

\item{natural.mortality}{nm}

\item{prop.f}{...}

\item{prop.m}{...}

\item{land.frac}{...}

\item{recapture}{...}

\item{sum.residual.fleets}{...}

\item{keep.all.ages}{...}
}
\description{
Combine the data sources to SAM readable object
}
\details{
...
}
