% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.sam}
\alias{simulate.sam}
\title{Simulate from a sam object}
\usage{
\method{simulate}{sam}(object, nsim = 1, seed = NULL, full.data = TRUE,
  ...)
}
\arguments{
\item{object}{sam fitted object (result from sam.fit)}

\item{nsim}{number of response lists to simulate. Defaults to 1.}

\item{seed}{random number seed}

\item{full.data}{logical, should each inner list contain a full list of data. Defaults to TRUE}

\item{...}{extra arguments}
}
\value{
returns a list of lists. The outer list has length \code{nsim}. Each inner list contains simulated values of \code{logF}, \code{logN}, and \code{obs} with dimensions equal to those parameters.
}
\description{
Simulate from a sam object
}
\details{
...
}

