% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{srplot}
\alias{srplot}
\title{Plots the stock recruitment}
\usage{
srplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments to plot}
}
\description{
Plots the stock recruitment
}
