% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{srplot}
\alias{srplot}
\alias{srplot.sam}
\title{Plots the stock recruitment}
\usage{
srplot(fit, ...)

\method{srplot}{sam}(
  fit,
  textcol = "red",
  years = TRUE,
  linetype = "l",
  linecol = "black",
  xlim,
  ylim,
  add = FALSE,
  CIlevel = 0.95,
  ...
)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments to plot}

\item{textcol}{color of years on plot}

\item{years}{the plotting symbols are the years}

\item{linetype}{type for the plot (default line)}

\item{linecol}{color of lines between points}

\item{xlim}{bounds for x-axis}

\item{ylim}{bounds for y-axis}

\item{add}{false if a new plot should be created}

\item{CIlevel}{Confidence level for error ellipseson stock-recruitment pairs}
}
\description{
Plots the stock recruitment
}
