% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssbplot}
\alias{ssbplot}
\alias{ssbplot.default}
\alias{ssbplot.samforecast}
\title{SAM SSB plot}
\usage{
ssbplot(fit, ...)

\method{ssbplot}{default}(fit, ...)

\method{ssbplot}{samforecast}(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM SSB plot
}
\details{
Plot of spawning stock biomass
}
