% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{ssbplot}
\alias{ssbplot}
\title{SAM SSB plot}
\usage{
ssbplot(fit, ...)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{...}{extra arguments transferred to plot including the following:
\code{add} logical, plotting is to be added on existing plot
\code{ci} logical, confidence intervals should be plotted
\code{cicol} color to plot the confidence polygon}
}
\description{
SAM SSB plot
}
\details{
...
}

