% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_referencepoints.R
\name{stochasticReferencepoints}
\alias{stochasticReferencepoints}
\alias{stochasticReferencepoints.sam}
\title{Estimate stochastic reference points}
\usage{
stochasticReferencepoints(fit, referencepoints, ...)

\method{stochasticReferencepoints}{sam}(
  fit,
  referencepoints,
  method = "Q0.5",
  catchType = "catch",
  nYears = 300,
  Frange = c(0, 2),
  nosim = 1000,
  aveYears = max(fit$data$years) + (-9:0),
  selYears = max(fit$data$years),
  newton.control = list(),
  seed = .timeToSeed(),
  formula = ~ibc(F, 5),
  nosim_ci = 200,
  derivedSummarizer = median,
  ...
)
}
\arguments{
\item{fit}{a sam fit}

\item{referencepoints}{a character vector of reference points to estimate (see Details)}

\item{...}{additional parameters that can be passed on}

\item{method}{estimation method (See Details)}

\item{catchType}{catch type: catch, landing, discard}

\item{nYears}{Number of years to forecast}

\item{Frange}{Range of F values to consider}

\item{nosim}{Number of simulations for estimation}

\item{aveYears}{Years to average over for biological input}

\item{selYears}{Years to average over for selectivity}

\item{newton.control}{List of control parameters for optimization}

\item{seed}{Seed for simulations}

\item{formula}{Formula to estimate optimization criteria as a function of F}

\item{nosim_ci}{Number of simulations for bootstrap confidence intervals}

\item{derivedSummarizer}{Function to summarize derived per-recruit values}
}
\value{
sam reference point object

reference point object
}
\description{
The function estimates reference points based on stochastic model forecasts.
The following reference points are implemented:
\describe{
   \item{F=x}{F fixed to x, e.g., \code{"F=0.3"} (NOT IMPLEMENTED YET)}
   \item{StatusQuo}{F in the last year of the assessment (NOT IMPLEMENTED YET)}
   \item{StatusQuo-y}{F in the y years before the last in the assessment, e.g., \code{"StatusQuo-1"} (NOT IMPLEMENTED YET)}
   \item{MSY}{F that maximizes yield}
   \item{0.xMSY}{Fs that gives 0.x*100\% of MSY, e.g., \code{"0.95MSY"}}
   \item{Max}{F that maximizes yield per recruit}
   \item{0.xdYPR}{F such that the derivative of yield per recruit is 0.x times the derivative at F=0, e.g., \code{"0.1dYPR"}}
   \item{0.xSPR}{F such that spawners per recruit is 0.x times spawners per recruit at F=0, e.g., \code{"0.35SPR"}}
   \item{0.xB0}{F such that biomass is 0.x times the biomass at F=0, e.g., \code{"0.2B0"}}
}
}
\details{
Reference points can be estimated using these methods:

\describe{
   \item{Mean}{Use least squares to estimate mean equilibrium values}
   \item{Q0.x}{Use quantile regression to estimate the 0.x quantile of equilibrium values}
   \item{Median}{Identical to Q0.5}
   \item{Mode}{(NOT IMPLEMENTED YET)}
}

To estimate median equilibrium yield, as required by ICES, the method "Q0.5" should be used.
}
\examples{
\dontrun{
  stochasticReferencepoints(fit, c("MSY","0.95MSY","Max","0.35SPR","0.1dYPR"))
}
}
