% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.sam}
\alias{summary.sam}
\title{Summary of sam object}
\usage{
\method{summary}{sam}(object, ...)
}
\arguments{
\item{object}{sam fitted object as returned from the \code{\link{sam.fit}} function}

\item{...}{extra arguments}
}
\description{
Summary of sam object
}
\details{
summary table containing recruits, SSB, and Fbar
}
