% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.data.files}
\alias{write.data.files}
\title{Write all data files from a list as created by 'setup.sam.data'}
\usage{
write.data.files(dat, dir = ".", writeToOne = TRUE, ...)
}
\arguments{
\item{dat}{A list as created by 'setup.sam.data'}

\item{dir}{Directory where the files are written}

\item{writeToOne}{Write multi fleet data to one file if data is equal for all fleets}

\item{...}{other arguments passes to write.ices}
}
\description{
Write all data files from a list as created by 'setup.sam.data'
}
\details{
Write all data files from a list as created by 'setup.sam.data'
}
