% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ypr}
\alias{ypr}
\title{Yield per recruit calculation}
\usage{
ypr(fit, Flimit = 2, Fdelta = 0.01, aveYears = 15, ageLimit = 100)
}
\arguments{
\item{fit}{the object returned from sam.fit}

\item{Flimit}{Upper limit for Fbar}

\item{Fdelta}{increments on the Fbar axis}

\item{aveYears}{Number of years back to use when calculating averages (selection, weights, ...)}

\item{ageLimit}{Oldest age used (should be high)}
}
\description{
Yield per recruit calculation
}
